refactor(jazz-next): separate runtime semantic layers
