+++ jazz-next/src/JazzNext/Compiler/Runtime.hs
@@ -33,7 +33,7 @@ module JazzNext.Compiler.Runtime
     untypedIntMetadata
   ) where
 
-import Control.Monad (foldM, zipWithM)
+import Control.Monad (foldM)
 import Control.Monad.Trans.Except
   ( ExceptT (..),
     runExceptT,
@@ -46,10 +46,9 @@ import Control.Monad.Trans.State.Strict
     modify',
     put
   )
-import Data.Char (chr, isAlpha, isAlphaNum, isControl, isDigit, isHexDigit, isSpace, ord, toUpper)
 import Data.Functor.Identity (runIdentity)
 import Data.List (foldl', scanl')
-import Data.Maybe (fromMaybe, isJust, listToMaybe)
+import Data.Maybe (fromMaybe)
 import qualified Data.IntMap.Lazy as LazyIntMap
 import qualified Data.Map.Lazy as LazyMap
 import qualified Data.Map.Strict as Map
@@ -61,47 +60,28 @@ import qualified Data.Text as Text
 import JazzNext.Compiler.AST
   ( CaseArm (..),
     ClassMethodSignature (..),
-    DataConstructorArgument (..),
     DataConstructor (..),
     Expr (..),
     ImplMethod (..),
     Literal (..),
     NumericType (..),
-    Pattern (..),
     SignaturePayload (..),
     SignatureType (..),
     Statement (..)
   )
 import JazzNext.Compiler.Diagnostics
   ( Diagnostic,
-    SourceSpan,
-    mkDiagnostic
-  )
-import JazzNext.Compiler.FractionalLiteral
-  ( FractionalLiteralSource,
-    fractionalLiteralExceedsMagnitude,
-    fractionalLiteralIntegralValue
+    SourceSpan
   )
 import JazzNext.Compiler.BuiltinCatalog
   ( BuiltinResolutionMode (..),
     BuiltinSymbol (..),
     builtinSymbolArity,
     builtinSymbolName,
-    builtinSymbolNumericConversionTarget,
-    lookupBuiltinSymbolInMode,
-    numericTypeFromName,
-    numericTypeFloatMax,
-    numericTypeIntegerBounds,
-    numericTypeIsIntegral,
-    renderNumericTypeName
+    lookupBuiltinSymbolInMode
   )
 import JazzNext.Compiler.CapabilityFacts
   ( concreteConstraintArgument,
-    constraintFunctionArgumentTypes,
-    constraintSignatureTypeContainsClassParameter,
-    constraintSignatureTypeVariableNamesInOrder,
-    constraintSignatureTypesCompatible,
-    identifierLooksLikeTypeVariable,
     qualifiedMethodKey,
     signaturePayloadConstraintType,
     substituteClassMethodSignature
@@ -109,7 +89,6 @@ import JazzNext.Compiler.CapabilityFacts
 import JazzNext.Compiler.Name
   ( GeneratedNameKind (..),
     Name (..),
-    NameNamespace (ConstructorNamespace),
     ResolvedNameOrigin (..),
     generatedName,
     identifierText,
@@ -122,9 +101,12 @@ import JazzNext.Compiler.Parser.Operator
 import JazzNext.Compiler.Pattern
   ( patternBinderNames
   )
+import JazzNext.Compiler.Runtime.Primitives
+  ( evalBinary,
+    evalBuiltin
+  )
 import JazzNext.Compiler.Runtime.ScopePlan
   ( buildRuntimeScopePlan,
-    exprContainsFunctionBranch,
     exprDefinitelyNotFunctionValue,
     runtimeExprRequiresHost,
     runtimeModulePathAfterStatements,
@@ -139,6 +121,34 @@ import JazzNext.Compiler.Runtime.ScopePlan
     scopePlanRecursiveGroupAt,
     scopePlanStatementAt
   )
+import JazzNext.Compiler.Runtime.Semantics
+  ( applyConstructor,
+    applyExplicitTypeApplicationResultHint,
+    applyRuntimeFunctionArgumentHint,
+    applyRuntimeFunctionResultHint,
+    applyRuntimeTypeHint,
+    attachDefaultBindingIntegerTarget,
+    attachRuntimeTypeHint,
+    convertFloatToNumericTarget,
+    convertIntegerToNumericTarget,
+    evalNumericConversion,
+    explicitTypeApplicationRuntimeFunctionHint,
+    explicitTypeApplicationRuntimeValueHint,
+    isFunctionValue,
+    literalRuntimeValue,
+    matchCaseArm,
+    numericConversionBuiltinForTarget,
+    preferredRuntimeMethodCandidates,
+    renderRuntimeType,
+    renderRuntimeValue,
+    runtimeConstraintType,
+    runtimeConstructorArgument,
+    runtimeDefinitionName,
+    runtimeDiagnostic,
+    runtimeQualifiedMethodIsFullyApplied,
+    runtimeValueExactlyMatchesConstraint,
+    untypedIntMetadata
+  )
 import JazzNext.Compiler.Runtime.Types
   ( DeferredHostBindingKey (..),
     DeferredHostBindingState (..),
@@ -147,10 +157,8 @@ import JazzNext.Compiler.Runtime.Types
     RuntimeCell,
     RuntimeEnv,
     RuntimeEvidence (..),
-    RuntimeFloatMetadata (..),
     RuntimeHostEvaluationState (..),
     RuntimeHostEvaluationT,
-    RuntimeIntMetadata (..),
     RuntimeMethodCandidate (..),
     RuntimeExplicitResultHints,
     RuntimeValue (..),
@@ -159,8 +167,7 @@ import JazzNext.Compiler.Runtime.Types
     foldRuntimeExplicitResultHints,
     pattern VExplicitResultHints,
     prependRuntimeExplicitResultHint,
-    runtimeExplicitResultHintsInOrder,
-    runtimeEvidenceTarget
+    runtimeExplicitResultHintsInOrder
   )
 import JazzNext.Compiler.RuntimeHints
   ( BindingRuntimeHintKey,
@@ -174,7 +181,6 @@ import JazzNext.Compiler.RuntimeHost
     hostIOCategoryToken,
     hostIOFailureMessage
   )
-import Numeric (showHex)
 
 runRuntimeHostEvaluation ::
   Monad m =>
@@ -337,107 +343,6 @@ evaluateRuntimeExprPureWithBuiltinsAndBindingHintsAndSourceUnitStatements prelud
           statements
     _ -> Just <$> evalValue builtinMode bindingTypeHints Map.empty expr
 
-renderRuntimeValue :: RuntimeValue -> Text
-renderRuntimeValue value =
-  case value of
-    VInt intValue _ -> Text.pack (show intValue)
-    VFloat floatValue _ -> Text.pack (show floatValue)
-    VBool boolValue ->
-      if boolValue
-        then "True"
-        else "False"
-    VChar charValue ->
-      "'" <> renderQuotedScalar charValue <> "'"
-    VText textValue ->
-      "\"" <> Text.concatMap renderQuotedScalar textValue <> "\""
-    VList elements _ ->
-      "[" <> Text.intercalate ", " (map renderRuntimeValue elements) <> "]"
-    VTuple elements ->
-      "(" <> Text.intercalate ", " (map renderRuntimeValue elements) <> ")"
-    VClosure {} -> "<function>"
-    VBuiltin _ _ -> "<function>"
-    VOperator {} -> "<function>"
-    VSectionLeft {} -> "<function>"
-    VSectionRight {} -> "<function>"
-    VConstructor _ _ constructorName constructorArguments capturedArgs
-      | constructorIsSaturated constructorArguments capturedArgs ->
-          renderConstructorValue constructorName capturedArgs
-      | otherwise ->
-          "<function>"
-    VQualifiedMethod {} -> "<function>"
-    VTyped _ innerValue -> renderRuntimeValue innerValue
-    VExplicitTypeApplication _ innerValue -> renderRuntimeValue innerValue
-    VExplicitResultHints _ innerValue -> renderRuntimeValue innerValue
-    VDeferredHostBinding {} -> "<deferred-host-binding>"
-
-renderQuotedScalar :: Char -> Text
-renderQuotedScalar value =
-  case value of
-    '\\' -> "\\\\"
-    '\'' -> "\\'"
-    '"' -> "\\\""
-    '\n' -> "\\n"
-    '\r' -> "\\r"
-    '\t' -> "\\t"
-    '\0' -> "\\0"
-    _
-      | isControl value ->
-          "\\u{" <> Text.pack (map toUpper (showHex (ord value) "")) <> "}"
-    _ -> Text.singleton value
-
-renderConstructorValue :: Name -> [RuntimeValue] -> Text
-renderConstructorValue constructorName arguments =
-  case arguments of
-    [] -> renderConstructorName constructorName
-    _ ->
-      renderConstructorName constructorName
-        <> "("
-        <> Text.intercalate ", " (map renderRuntimeValue arguments)
-        <> ")"
-
-renderConstructorName :: Name -> Text
-renderConstructorName constructorName =
-  case constructorName of
-    ResolvedName _ ConstructorNamespace identifier -> identifierText identifier
-    _ -> identifierText constructorName
-
-runtimeDefinitionName :: Maybe [Text] -> Name -> Name
-runtimeDefinitionName maybeModulePath name =
-  case (maybeModulePath, name) of
-    (Just modulePath, ResolvedName CurrentModule namespace identifier) ->
-      ResolvedName (ImportedModule modulePath) namespace identifier
-    _ -> name
-
-runtimeConstructorArgument :: Maybe [Text] -> DataConstructorArgument -> DataConstructorArgument
-runtimeConstructorArgument maybeModulePath argument =
-  case argument of
-    DataConstructorArgumentName name ->
-      DataConstructorArgumentName (runtimeTypeName maybeModulePath name)
-    DataConstructorArgumentOpaque -> DataConstructorArgumentOpaque
-
-runtimeConstraintType :: Maybe [Text] -> SignatureType -> SignatureType
-runtimeConstraintType maybeModulePath signatureType =
-  case signatureType of
-    TypeVariable name -> TypeVariable (runtimeTypeName maybeModulePath name)
-    TypeName name -> TypeName (runtimeTypeName maybeModulePath name)
-    TypeApplication name arguments ->
-      TypeApplication
-        (runtimeTypeName maybeModulePath name)
-        (map (runtimeConstraintType maybeModulePath) arguments)
-    TypeList elementType -> TypeList (runtimeConstraintType maybeModulePath elementType)
-    TypeTuple elementTypes -> TypeTuple (map (runtimeConstraintType maybeModulePath) elementTypes)
-    TypeFunction argumentType resultType ->
-      TypeFunction
-        (runtimeConstraintType maybeModulePath argumentType)
-        (runtimeConstraintType maybeModulePath resultType)
-    _ -> signatureType
-
-runtimeTypeName :: Maybe [Text] -> Name -> Name
-runtimeTypeName maybeModulePath name
-  | identifierText name `elem` ["Int", "Float", "Bool", "Char", "Text"] = name
-  | Just _ <- numericTypeFromName (identifierText name) = name
-  | identifierLooksLikeTypeVariable name = name
-  | otherwise = runtimeDefinitionName maybeModulePath name
 
 -- Public scope entry points receive an opaque map whose lazy cells may include
 -- recursive blackholes. They cannot safely recover provenance by inspecting
@@ -1390,1951 +1295,6 @@ declaredOperatorRightSectionClosure currentModulePath operatorValue rightValue e
       Map.insert functionName (Right operatorValue) $
         Map.insert rightParameter (Right rightValue) env
 
-literalRuntimeValue :: Literal -> RuntimeValue
-literalRuntimeValue literal =
-  case literal of
-    LInt value -> VInt value untypedIntMetadata
-    LFloat value literalSource maybeTargetType ->
-      case maybeTargetType of
-        Just targetType ->
-          VFloat
-            (roundFloatTarget targetType value)
-            (targetedFloatMetadataWithSource targetType (Just literalSource))
-        Nothing ->
-          VFloat value (untypedFloatMetadata (Just literalSource))
-    LBool value -> VBool value
-    LChar value -> VChar value
-    LText value -> VText value
-
-attachRuntimeTypeHint :: Maybe SignatureType -> RuntimeValue -> Either Diagnostic RuntimeValue
-attachRuntimeTypeHint maybeTypeHint runtimeValue =
-  case maybeTypeHint of
-    Just typeHint ->
-      applyRuntimeTypeHint typeHint runtimeValue
-    Nothing ->
-      Right runtimeValue
-
-applyRuntimeTypeHint :: SignatureType -> RuntimeValue -> Either Diagnostic RuntimeValue
-applyRuntimeTypeHint typeHint runtimeValue =
-  case runtimeValue of
-    VTyped existingTypeHint _
-      | runtimeTypeHintAtLeastAsSpecific existingTypeHint typeHint ->
-          Right runtimeValue
-    VTyped _ innerValue ->
-      applyRuntimeTypeHint typeHint innerValue
-    VExplicitTypeApplication _ innerValue ->
-      applyRuntimeTypeHint typeHint innerValue
-    VExplicitResultHints _ innerValue ->
-      applyRuntimeTypeHint typeHint innerValue
-    _ ->
-      case (typeHint, runtimeValue) of
-        (TypeInt, _) -> do
-          convertedValue <- evalNumericConversion (numericConversionBuiltinForTarget NumericInt64) NumericInt64 runtimeValue
-          Right (VTyped TypeInt convertedValue)
-        (TypeFloat, _) -> do
-          convertedValue <- evalNumericConversion (numericConversionBuiltinForTarget NumericFloat64) NumericFloat64 runtimeValue
-          Right (VTyped TypeFloat convertedValue)
-        (TypeNumeric targetType, _) ->
-          evalNumericConversion (numericConversionBuiltinForTarget targetType) targetType runtimeValue
-        (TypeBool, VBool {}) -> Right runtimeValue
-        (TypeChar, VChar {}) -> Right runtimeValue
-        (TypeText, VText {}) -> Right runtimeValue
-        (TypeName typeName, _)
-          | Just targetType <- constraintTypeNameNumericTarget typeName -> do
-              convertedValue <- evalNumericConversion (numericConversionBuiltinForTarget targetType) targetType runtimeValue
-              if identifierText typeName == "Int" || identifierText typeName == "Float"
-                then Right (VTyped typeHint convertedValue)
-                else Right convertedValue
-        (TypeName typeName, VChar {})
-          | identifierText typeName == "Char" ->
-              Right runtimeValue
-        (TypeName typeName, VText {})
-          | identifierText typeName == "Text" ->
-              Right runtimeValue
-        (TypeName hintedTypeName, VConstructor typeName typeParameters constructorName constructorArguments capturedArgs)
-          | identifierText hintedTypeName == identifierText typeName,
-            constructorIsSaturated constructorArguments capturedArgs -> do
-              hintedCapturedArgs <-
-                zipWithM
-                  (applyConstructorArgumentRuntimeHint Map.empty)
-                  constructorArguments
-                  capturedArgs
-              Right
-                ( VTyped
-                    typeHint
-                    (VConstructor typeName typeParameters constructorName constructorArguments hintedCapturedArgs)
-                )
-        (TypeList _, VList _ (Just existingTypeHint))
-          | runtimeTypeHintAtLeastAsSpecific existingTypeHint typeHint ->
-              Right runtimeValue
-        (TypeList elementType, VList elements _) -> do
-          hintedElements <- mapM (applyRuntimeTypeHint elementType) elements
-          Right (VList hintedElements (Just typeHint))
-        (TypeTuple elementTypes, VTuple elements)
-          | length elementTypes == length elements ->
-              VTuple <$> zipWithM applyRuntimeTypeHint elementTypes elements
-        (TypeFunction {}, VClosure capturedEnv capturedEnvMayReachHostCells parameterName bodyExpr _ closureModulePath) ->
-          Right (VClosure capturedEnv capturedEnvMayReachHostCells parameterName bodyExpr (Just typeHint) closureModulePath)
-        (TypeFunction {}, _)
-          | isFunctionValue runtimeValue ->
-              Right (VTyped typeHint runtimeValue)
-        (TypeApplication hintedTypeName hintedArguments, VConstructor typeName typeParameters constructorName constructorArguments capturedArgs)
-          | identifierText hintedTypeName == identifierText typeName,
-            length hintedArguments == length typeParameters -> do
-              let typeParameterHints =
-                    Map.fromList (zip (map identifierText typeParameters) hintedArguments)
-              hintedCapturedArgs <-
-                zipWithM
-                  (applyConstructorArgumentRuntimeHint typeParameterHints)
-                  constructorArguments
-                  capturedArgs
-              Right (VTyped typeHint (VConstructor typeName typeParameters constructorName constructorArguments hintedCapturedArgs))
-        _ ->
-          Right runtimeValue
-
--- Runtime hints form an information order rather than a replacement order.
--- A concrete value already carrying, for example, @[CanonicalToken]@ also
--- satisfies a later polymorphic @[a]@ result hint. Preserving the stronger
--- evidence avoids both losing concrete dispatch information and repeatedly
--- traversing persistent values at polymorphic function boundaries.
-runtimeTypeHintAtLeastAsSpecific :: SignatureType -> SignatureType -> Bool
-runtimeTypeHintAtLeastAsSpecific existingHint requestedHint
-  | existingHint == requestedHint = True
-runtimeTypeHintAtLeastAsSpecific _ (TypeVariable _) = True
-runtimeTypeHintAtLeastAsSpecific _ (TypeName name)
-  | identifierLooksLikeTypeVariable name = True
-runtimeTypeHintAtLeastAsSpecific
-  (TypeApplication existingName existingArguments)
-  (TypeApplication requestedName requestedArguments) =
-    existingName == requestedName
-    && length existingArguments == length requestedArguments
-    && and (zipWith runtimeTypeHintAtLeastAsSpecific existingArguments requestedArguments)
-runtimeTypeHintAtLeastAsSpecific (TypeList existingElement) (TypeList requestedElement) =
-  runtimeTypeHintAtLeastAsSpecific existingElement requestedElement
-runtimeTypeHintAtLeastAsSpecific (TypeTuple existingElements) (TypeTuple requestedElements) =
-  length existingElements == length requestedElements
-    && and (zipWith runtimeTypeHintAtLeastAsSpecific existingElements requestedElements)
-runtimeTypeHintAtLeastAsSpecific
-  (TypeFunction existingArgument existingResult)
-  (TypeFunction requestedArgument requestedResult) =
-    runtimeTypeHintAtLeastAsSpecific existingArgument requestedArgument
-      && runtimeTypeHintAtLeastAsSpecific existingResult requestedResult
-runtimeTypeHintAtLeastAsSpecific _ _ = False
-
-applyConstructorArgumentRuntimeHint ::
-  Map Text SignatureType ->
-  DataConstructorArgument ->
-  RuntimeValue ->
-  Either Diagnostic RuntimeValue
-applyConstructorArgumentRuntimeHint typeParameterHints constructorArgument runtimeValue =
-  case constructorArgument of
-    DataConstructorArgumentName argumentName ->
-      attachRuntimeTypeHint (constructorArgumentRuntimeHint typeParameterHints argumentName) runtimeValue
-    DataConstructorArgumentOpaque ->
-      Right runtimeValue
-
-constructorArgumentRuntimeHint :: Map Text SignatureType -> Name -> Maybe SignatureType
-constructorArgumentRuntimeHint typeParameterHints argumentName =
-  case Map.lookup (identifierText argumentName) typeParameterHints of
-    Just hintedType -> Just hintedType
-    Nothing -> concreteConstructorPayloadRuntimeHint argumentName
-
-concreteConstructorPayloadRuntimeHint :: Name -> Maybe SignatureType
-concreteConstructorPayloadRuntimeHint argumentName
-  | identifierText argumentName == "Int" = Just TypeInt
-  | identifierText argumentName == "Float" = Just TypeFloat
-  | Just numericType <- constraintTypeNameNumericTarget argumentName =
-      Just (TypeNumeric numericType)
-  | identifierText argumentName == "Bool" =
-      Just TypeBool
-  | identifierText argumentName == "Char" = Just TypeChar
-  | identifierText argumentName == "Text" = Just TypeText
-  | otherwise =
-      Nothing
-
-constraintTypeNameNumericTarget :: Name -> Maybe NumericType
-constraintTypeNameNumericTarget typeName =
-  case identifierText typeName of
-    "Int" -> Just NumericInt64
-    "Int8" -> Just NumericInt8
-    "Int16" -> Just NumericInt16
-    "Int32" -> Just NumericInt32
-    "Int64" -> Just NumericInt64
-    "UInt8" -> Just NumericUInt8
-    "UInt16" -> Just NumericUInt16
-    "UInt32" -> Just NumericUInt32
-    "UInt64" -> Just NumericUInt64
-    "Float" -> Just NumericFloat64
-    "Float16" -> Just NumericFloat16
-    "Float32" -> Just NumericFloat32
-    "Float64" -> Just NumericFloat64
-    _ -> Nothing
-
-untypedIntMetadata :: RuntimeIntMetadata
-untypedIntMetadata =
-  RuntimeIntMetadata {runtimeIntTargetType = Nothing}
-
-targetedIntMetadata :: NumericType -> RuntimeIntMetadata
-targetedIntMetadata targetType =
-  RuntimeIntMetadata {runtimeIntTargetType = Just targetType}
-
-untypedFloatMetadata :: Maybe FractionalLiteralSource -> RuntimeFloatMetadata
-untypedFloatMetadata literalSource =
-  RuntimeFloatMetadata
-    { runtimeFloatLiteralSource = literalSource,
-      runtimeFloatTargetType = Nothing
-    }
-
-targetedFloatMetadata :: NumericType -> RuntimeFloatMetadata
-targetedFloatMetadata targetType =
-  targetedFloatMetadataWithSource targetType Nothing
-
-targetedFloatMetadataWithSource :: NumericType -> Maybe FractionalLiteralSource -> RuntimeFloatMetadata
-targetedFloatMetadataWithSource targetType literalSource =
-  RuntimeFloatMetadata
-    { runtimeFloatLiteralSource =
-        case targetType of
-          NumericFloat64 -> literalSource
-          _ -> Nothing,
-      runtimeFloatTargetType = Just targetType
-    }
-
--- | Pattern bindings are prepended to the arm environment so they shadow outer
--- runtime bindings only while evaluating the selected arm body.
-matchCaseArm ::
-  Maybe [Text] ->
-  RuntimeEnv ->
-  RuntimeValue ->
-  CaseArm ->
-  Maybe (RuntimeEnv, Maybe Expr, Expr)
-matchCaseArm currentModulePath env scrutineeValue (CaseArm casePattern guardExpr bodyExpr) =
-  case matchPattern currentModulePath scrutineeValue casePattern of
-    Just patternBindings ->
-      Just (Map.union patternBindings env, guardExpr, bodyExpr)
-    Nothing -> Nothing
-
-matchPattern :: Maybe [Text] -> RuntimeValue -> Pattern -> Maybe RuntimeEnv
-matchPattern currentModulePath scrutineeValue casePattern =
-  case casePattern of
-    PWildcard -> Just Map.empty
-    PVariable name ->
-      Just
-        (Map.singleton name (Right scrutineeValue))
-    PLiteral literal
-      | scrutineeValue == literalRuntimeValue literal ->
-          Just Map.empty
-      | otherwise ->
-          Nothing
-    PConstructor constructorName patterns ->
-      case constructorPatternScrutinee scrutineeValue of
-        VConstructor _ _ valueConstructorName constructorArguments capturedArgs
-          | valueConstructorName == runtimeDefinitionName currentModulePath constructorName,
-            constructorIsSaturated constructorArguments capturedArgs,
-            length capturedArgs == length patterns ->
-              matchPatternList currentModulePath capturedArgs patterns
-        _ -> Nothing
-    PList patterns ->
-      case scrutineeValue of
-        VList elements _
-          | length elements == length patterns ->
-              matchPatternList currentModulePath elements patterns
-        _ -> Nothing
-    PConsList headPattern tailPattern ->
-      case scrutineeValue of
-        VList (headValue : tailValues) maybeTypeHint -> do
-          headBindings <- matchPattern currentModulePath headValue headPattern
-          tailBindings <- matchPattern currentModulePath (VList tailValues maybeTypeHint) tailPattern
-          Just (tailBindings `Map.union` headBindings)
-        _ -> Nothing
-    PTuple patterns ->
-      case scrutineeValue of
-        VTuple elements
-          | length elements == length patterns ->
-              matchPatternList currentModulePath elements patterns
-        _ -> Nothing
-    PAs name nestedPattern -> do
-      patternBindings <- matchPattern currentModulePath scrutineeValue nestedPattern
-      Just (Map.insert name (Right scrutineeValue) patternBindings)
-    POr alternatives ->
-      matchFirstAlternative currentModulePath scrutineeValue alternatives
-
-matchFirstAlternative :: Maybe [Text] -> RuntimeValue -> [Pattern] -> Maybe RuntimeEnv
-matchFirstAlternative currentModulePath scrutineeValue alternatives =
-  case alternatives of
-    [] -> Nothing
-    alternative : rest ->
-      case matchPattern currentModulePath scrutineeValue alternative of
-        Just patternBindings -> Just patternBindings
-        Nothing -> matchFirstAlternative currentModulePath scrutineeValue rest
-
-matchPatternList :: Maybe [Text] -> [RuntimeValue] -> [Pattern] -> Maybe RuntimeEnv
-matchPatternList currentModulePath values patterns =
-  foldM step Map.empty (zip values patterns)
-  where
-    step bindings (value, elementPattern) =
-      case matchPattern currentModulePath value elementPattern of
-        Just patternBindings -> Just (patternBindings `Map.union` bindings)
-        Nothing -> Nothing
-
-constructorPatternScrutinee :: RuntimeValue -> RuntimeValue
-constructorPatternScrutinee runtimeValue =
-  case runtimeValue of
-    VTyped _ innerValue -> constructorPatternScrutinee innerValue
-    VExplicitTypeApplication _ innerValue -> constructorPatternScrutinee innerValue
-    VExplicitResultHints _ innerValue -> constructorPatternScrutinee innerValue
-    _ -> runtimeValue
-
--- | Apply any callable runtime value, including sections, builtin primitives,
--- and curried operator values.
-applyRuntimeFunction ::
-  BuiltinResolutionMode ->
-  Map BindingRuntimeHintKey SignatureType ->
-  RuntimeValue ->
-  RuntimeValue ->
-  Either Diagnostic RuntimeValue
-applyRuntimeFunction builtinMode bindingTypeHints functionValue argumentValue =
-  runIdentity
-    ( runRuntimeHostEvaluation disabledRuntimeHost $ \host ->
-        runExceptT
-          ( runCallableMachine
-              host
-              builtinMode
-              bindingTypeHints
-              functionValue
-              argumentValue
-          )
-    )
-
-applyRuntimeFunctionResultHint :: SignatureType -> RuntimeValue -> Either Diagnostic RuntimeValue
-applyRuntimeFunctionResultHint typeHint runtimeValue =
-  case typeHint of
-    TypeFunction _ resultType ->
-      applyRuntimeTypeHint resultType runtimeValue
-    _ ->
-      Right runtimeValue
-
-applyRuntimeFunctionArgumentHint :: SignatureType -> RuntimeValue -> Either Diagnostic RuntimeValue
-applyRuntimeFunctionArgumentHint typeHint runtimeValue =
-  case typeHint of
-    TypeFunction argumentType _ ->
-      applyRuntimeTypeHint argumentType runtimeValue
-    _ ->
-      Right runtimeValue
-
-applyExplicitTypeApplicationResultHint :: SignatureType -> RuntimeValue -> Either Diagnostic RuntimeValue
-applyExplicitTypeApplicationResultHint typeHint runtimeValue
-  | isFunctionValue runtimeValue =
-      Right (prependRuntimeExplicitResultHint typeHint runtimeValue)
-  | runtimeValueCanAcceptTypeHint typeHint runtimeValue =
-      applyRuntimeTypeHint typeHint runtimeValue
-  | otherwise =
-      Right runtimeValue
-
-runtimeValueCanAcceptTypeHint :: SignatureType -> RuntimeValue -> Bool
-runtimeValueCanAcceptTypeHint typeHint runtimeValue =
-  case runtimeValue of
-    VTyped _ innerValue ->
-      runtimeValueCanAcceptTypeHint typeHint innerValue
-    VExplicitTypeApplication _ innerValue ->
-      runtimeValueCanAcceptTypeHint typeHint innerValue
-    VExplicitResultHints _ innerValue ->
-      runtimeValueCanAcceptTypeHint typeHint innerValue
-    _ ->
-      case (typeHint, runtimeValue) of
-        (TypeInt, VInt {}) -> True
-        (TypeFloat, VFloat {}) -> True
-        (TypeNumeric _, VInt {}) -> True
-        (TypeNumeric _, VFloat {}) -> True
-        (TypeBool, VBool {}) -> True
-        (TypeChar, VChar {}) -> True
-        (TypeText, VText {}) -> True
-        (TypeName typeName, VInt {}) ->
-          identifierText typeName == "Int" || isJust (constraintTypeNameNumericTarget typeName)
-        (TypeName typeName, VFloat {}) ->
-          identifierText typeName == "Float" || isJust (constraintTypeNameNumericTarget typeName)
-        (TypeName typeName, VBool {}) ->
-          identifierText typeName == "Bool"
-        (TypeName typeName, VChar {}) ->
-          identifierText typeName == "Char"
-        (TypeName typeName, VText {}) ->
-          identifierText typeName == "Text"
-        (TypeName typeName, VConstructor constructorTypeName _ _ constructorArguments capturedArgs) ->
-          identifierText typeName == identifierText constructorTypeName
-            && constructorIsSaturated constructorArguments capturedArgs
-        (TypeApplication typeName arguments, VConstructor constructorTypeName typeParameters _ constructorArguments capturedArgs) ->
-          identifierText typeName == identifierText constructorTypeName
-            && length arguments == length typeParameters
-            && constructorIsSaturated constructorArguments capturedArgs
-        (TypeList {}, VList {}) ->
-          True
-        (TypeTuple elementTypes, VTuple elements) ->
-          length elementTypes == length elements
-        (TypeFunction {}, _) ->
-          isFunctionValue runtimeValue
-        _ ->
-          False
-
-explicitTypeApplicationRuntimeFunctionHint :: SignatureType -> RuntimeValue -> Maybe SignatureType
-explicitTypeApplicationRuntimeFunctionHint typeHint runtimeValue = do
-  explicitTypeApplicationRuntimeTemplateHint typeHint runtimeValue
-
-explicitTypeApplicationRuntimeValueHint :: SignatureType -> RuntimeValue -> Maybe SignatureType
-explicitTypeApplicationRuntimeValueHint typeHint runtimeValue =
-  case explicitTypeApplicationRuntimeTemplateHint typeHint runtimeValue of
-    Just instantiatedTemplate -> Just instantiatedTemplate
-    Nothing -> explicitTypeApplicationRuntimeShapeHint typeHint runtimeValue
-
-explicitTypeApplicationRuntimeTemplateHint :: SignatureType -> RuntimeValue -> Maybe SignatureType
-explicitTypeApplicationRuntimeTemplateHint typeHint runtimeValue = do
-  templateHint <- runtimeValueSignatureHint runtimeValue
-  variableName <- listToMaybe (constraintSignatureTypeVariableNamesInOrder templateHint)
-  pure (substituteSignatureTypeVariable variableName typeHint templateHint)
-
-explicitTypeApplicationRuntimeShapeHint :: SignatureType -> RuntimeValue -> Maybe SignatureType
-explicitTypeApplicationRuntimeShapeHint typeHint runtimeValue =
-  case runtimeValue of
-    VTyped _ innerValue ->
-      explicitTypeApplicationRuntimeShapeHint typeHint innerValue
-    VExplicitTypeApplication _ innerValue ->
-      explicitTypeApplicationRuntimeShapeHint typeHint innerValue
-    VExplicitResultHints _ innerValue ->
-      explicitTypeApplicationRuntimeShapeHint typeHint innerValue
-    VList {} ->
-      Just (TypeList typeHint)
-    VConstructor typeName typeParameters _ constructorArguments capturedArgs
-      | length typeParameters == 1,
-        constructorIsSaturated constructorArguments capturedArgs ->
-          Just (TypeApplication typeName [typeHint])
-    _ -> Nothing
-
-runtimeValueSignatureHint :: RuntimeValue -> Maybe SignatureType
-runtimeValueSignatureHint runtimeValue =
-  case runtimeValue of
-    VTyped typeHint _ ->
-      Just typeHint
-    VExplicitTypeApplication _ innerValue ->
-      runtimeValueSignatureHint innerValue
-    VExplicitResultHints _ innerValue ->
-      runtimeValueSignatureHint innerValue
-    VClosure _ _ _ _ maybeTypeHint _ ->
-      maybeTypeHint
-    VList _ (Just typeHint) ->
-      Just typeHint
-    _ -> Nothing
-
-substituteSignatureTypeVariable :: Text -> SignatureType -> SignatureType -> SignatureType
-substituteSignatureTypeVariable variableName replacementType signatureType =
-  case signatureType of
-    TypeVariable name
-      | identifierText name == variableName -> replacementType
-      | otherwise -> signatureType
-    TypeName name
-      | identifierLooksLikeTypeVariable name,
-        identifierText name == variableName ->
-          replacementType
-      | otherwise ->
-          signatureType
-    TypeApplication typeName arguments ->
-      TypeApplication typeName (map (substituteSignatureTypeVariable variableName replacementType) arguments)
-    TypeList innerType ->
-      TypeList (substituteSignatureTypeVariable variableName replacementType innerType)
-    TypeTuple elementTypes ->
-      TypeTuple (map (substituteSignatureTypeVariable variableName replacementType) elementTypes)
-    TypeFunction argumentType resultType ->
-      TypeFunction
-        (substituteSignatureTypeVariable variableName replacementType argumentType)
-        (substituteSignatureTypeVariable variableName replacementType resultType)
-    _ -> signatureType
-
-runtimeQualifiedMethodIsFullyApplied ::
-  Text ->
-  SignaturePayload ->
-  [RuntimeValue] ->
-  [RuntimeMethodCandidate] ->
-  Bool
-runtimeQualifiedMethodIsFullyApplied classParameter methodSignature arguments candidates =
-  any candidateIsFullyApplied candidates
-  where
-    candidateIsFullyApplied (RuntimeMethodCandidate evidence _) =
-      case substituteClassMethodSignature classParameter implTarget methodSignature of
-        Just substitutedSignature ->
-          let (argumentTypes, _) = constraintFunctionArgumentTypes substitutedSignature
-           in length arguments >= length argumentTypes
-        Nothing ->
-          False
-      where
-        implTarget = runtimeEvidenceTarget evidence
-
-runtimeMethodCandidateExactlyMatches :: Text -> SignaturePayload -> [RuntimeValue] -> RuntimeMethodCandidate -> Bool
-runtimeMethodCandidateExactlyMatches classParameter methodSignature arguments (RuntimeMethodCandidate evidence _) =
-  case (signaturePayloadConstraintType methodSignature, substituteClassMethodSignature classParameter implTarget methodSignature) of
-    (Just genericSignature, Just substitutedSignature) ->
-      let (genericArgumentTypes, _) = constraintFunctionArgumentTypes genericSignature
-          (argumentTypes, _) = constraintFunctionArgumentTypes substitutedSignature
-          suppliedArgumentCount = length arguments
-          suppliedGenericArgumentTypes = take suppliedArgumentCount genericArgumentTypes
-          suppliedArgumentTypes = take suppliedArgumentCount argumentTypes
-          targetArgumentPositions =
-            map (constraintSignatureTypeContainsClassParameter classParameter) suppliedGenericArgumentTypes
-       in suppliedArgumentCount <= length genericArgumentTypes
-            && suppliedArgumentCount <= length argumentTypes
-            && or targetArgumentPositions
-            && and
-              ( zipWith3
-                  runtimeExactCandidateArgumentMatches
-                  targetArgumentPositions
-                  suppliedArgumentTypes
-                  arguments
-              )
-    _ ->
-      False
-  where
-    implTarget = runtimeEvidenceTarget evidence
-
-runtimeExactCandidateArgumentMatches :: Bool -> SignatureType -> RuntimeValue -> Bool
-runtimeExactCandidateArgumentMatches targetArgumentPosition signatureType runtimeValue =
-  not targetArgumentPosition || runtimeValueExactlyMatchesConstraint signatureType runtimeValue
-
-runtimeValueExactlyMatchesConstraint :: SignatureType -> RuntimeValue -> Bool
-runtimeValueExactlyMatchesConstraint signatureType runtimeValue =
-  case runtimeValue of
-    VExplicitTypeApplication _ innerValue ->
-      runtimeValueExactlyMatchesConstraint signatureType innerValue
-    VExplicitResultHints _ innerValue ->
-      runtimeValueExactlyMatchesConstraint signatureType innerValue
-    VTyped typeHint _ ->
-      typeHint == signatureType
-    VClosure _ _ _ _ (Just typeHint) _ ->
-      typeHint == signatureType
-    VInt _ metadata ->
-      case signatureType of
-        TypeInt -> runtimeIntTargetType metadata == Nothing
-        TypeNumeric numericType -> runtimeIntTargetType metadata == Just numericType
-        TypeName typeName ->
-          runtimeIntExactlyMatchesTypeName (identifierText typeName) metadata
-        _ -> False
-    VFloat _ metadata ->
-      case signatureType of
-        TypeFloat -> runtimeFloatTargetType metadata == Nothing
-        TypeNumeric numericType -> runtimeFloatTargetType metadata == Just numericType
-        TypeName typeName ->
-          runtimeFloatExactlyMatchesTypeName (identifierText typeName) metadata
-        _ -> False
-    VChar {} ->
-      case signatureType of
-        TypeChar -> True
-        TypeName typeName -> identifierText typeName == "Char"
-        _ -> False
-    VText {} ->
-      case signatureType of
-        TypeText -> True
-        TypeName typeName -> identifierText typeName == "Text"
-        _ -> False
-    VBool {} ->
-      case signatureType of
-        TypeBool -> True
-        TypeName typeName -> identifierText typeName == "Bool"
-        _ -> False
-    VList _ (Just typeHint) ->
-      typeHint == signatureType
-    VList elements Nothing ->
-      case signatureType of
-        TypeList elementType ->
-          not (null elements)
-            && all (runtimeValueExactlyMatchesConstraint elementType) elements
-        _ -> False
-    VTuple elements ->
-      case signatureType of
-        TypeTuple elementTypes
-          | length elementTypes == length elements ->
-              and (zipWith runtimeValueExactlyMatchesConstraint elementTypes elements)
-        _ -> False
-    VConstructor {} ->
-      case signatureType of
-        TypeName typeName ->
-          runtimeValueExactlyMatchesDataTypeName typeName runtimeValue
-        TypeApplication typeName typeArguments ->
-          runtimeValueExactlyMatchesDataTypeApplication typeName typeArguments runtimeValue
-        _ -> False
-    _ -> False
-
-runtimeIntExactlyMatchesTypeName :: Text -> RuntimeIntMetadata -> Bool
-runtimeIntExactlyMatchesTypeName typeName metadata =
-  case (typeName, runtimeIntTargetType metadata) of
-    ("Int", Nothing) -> True
-    ("Int8", Just NumericInt8) -> True
-    ("Int16", Just NumericInt16) -> True
-    ("Int32", Just NumericInt32) -> True
-    ("Int64", Just NumericInt64) -> True
-    ("UInt8", Just NumericUInt8) -> True
-    ("UInt16", Just NumericUInt16) -> True
-    ("UInt32", Just NumericUInt32) -> True
-    ("UInt64", Just NumericUInt64) -> True
-    _ -> False
-
-runtimeFloatExactlyMatchesTypeName :: Text -> RuntimeFloatMetadata -> Bool
-runtimeFloatExactlyMatchesTypeName typeName metadata =
-  case (typeName, runtimeFloatTargetType metadata) of
-    ("Float", Nothing) -> True
-    ("Float16", Just NumericFloat16) -> True
-    ("Float32", Just NumericFloat32) -> True
-    ("Float64", Just NumericFloat64) -> True
-    _ -> False
-
-runtimeMethodCandidateMatches :: Text -> SignaturePayload -> [RuntimeValue] -> RuntimeMethodCandidate -> Bool
-runtimeMethodCandidateMatches classParameter methodSignature arguments (RuntimeMethodCandidate evidence _) =
-  case substituteClassMethodSignature classParameter implTarget methodSignature of
-    Just substitutedSignature ->
-      let (argumentTypes, _) = constraintFunctionArgumentTypes substitutedSignature
-       in length arguments <= length argumentTypes
-            && and (zipWith runtimeValueMatchesConstraint argumentTypes arguments)
-    Nothing ->
-      False
-  where
-    implTarget = runtimeEvidenceTarget evidence
-
-runtimeValueMatchesConstraint :: SignatureType -> RuntimeValue -> Bool
-runtimeValueMatchesConstraint signatureType runtimeValue =
-  case runtimeValue of
-    VExplicitTypeApplication _ innerValue ->
-      runtimeValueMatchesConstraint signatureType innerValue
-    VExplicitResultHints _ innerValue ->
-      runtimeValueMatchesConstraint signatureType innerValue
-    VTyped typeHint _ ->
-      constraintSignatureTypesCompatible typeHint signatureType
-    _ ->
-      case signatureType of
-        TypeInt -> runtimeValueMatchesTypeName "Int" runtimeValue
-        TypeFloat -> runtimeValueMatchesTypeName "Float" runtimeValue
-        TypeNumeric numericType -> runtimeValueMatchesTypeName (renderNumericTypeName numericType) runtimeValue
-        TypeBool -> runtimeValueMatchesTypeName "Bool" runtimeValue
-        TypeChar -> runtimeValueMatchesTypeName "Char" runtimeValue
-        TypeText -> runtimeValueMatchesTypeName "Text" runtimeValue
-        TypeVariable {} -> False
-        TypeName typeName ->
-          runtimeValueMatchesTypeName (identifierText typeName) runtimeValue
-        TypeApplication typeName typeArguments ->
-          runtimeValueMatchesDataTypeApplication typeName typeArguments runtimeValue
-        TypeList elementType ->
-          case runtimeValue of
-            VList elements maybeTypeHint ->
-              case maybeTypeHint of
-                Just typeHint -> constraintSignatureTypesCompatible typeHint signatureType
-                Nothing -> all (runtimeValueMatchesConstraint elementType) elements
-            _ -> False
-        TypeTuple elementTypes ->
-          case runtimeValue of
-            VTuple elements
-              | length elementTypes == length elements ->
-                  and (zipWith runtimeValueMatchesConstraint elementTypes elements)
-            _ -> False
-        TypeFunction {} ->
-          case runtimeValue of
-            VClosure _ _ _ _ (Just typeHint) _ -> constraintSignatureTypesCompatible typeHint signatureType
-            _ -> isFunctionValue runtimeValue
-
-runtimeValueMatchesTypeName :: Text -> RuntimeValue -> Bool
-runtimeValueMatchesTypeName typeName runtimeValue =
-  case typeName of
-    "Int" -> runtimeIntMatchesIntAlias runtimeValue
-    "Int8" -> runtimeIntMatchesTarget NumericInt8 runtimeValue
-    "Int16" -> runtimeIntMatchesTarget NumericInt16 runtimeValue
-    "Int32" -> runtimeIntMatchesTarget NumericInt32 runtimeValue
-    "Int64" -> runtimeIntMatchesTarget NumericInt64 runtimeValue
-    "UInt8" -> runtimeIntMatchesTarget NumericUInt8 runtimeValue
-    "UInt16" -> runtimeIntMatchesTarget NumericUInt16 runtimeValue
-    "UInt32" -> runtimeIntMatchesTarget NumericUInt32 runtimeValue
-    "UInt64" -> runtimeIntMatchesTarget NumericUInt64 runtimeValue
-    "Float" -> runtimeFloatMatchesFloatAlias runtimeValue
-    "Float16" -> runtimeFloatHasTarget NumericFloat16 runtimeValue
-    "Float32" -> runtimeFloatHasTarget NumericFloat32 runtimeValue
-    "Float64" -> runtimeFloatHasTarget NumericFloat64 runtimeValue
-    "Bool" -> isRuntimeBool runtimeValue
-    "Char" -> isRuntimeChar runtimeValue
-    "Text" -> isRuntimeText runtimeValue
-    _ -> runtimeValueMatchesDataTypeName typeName runtimeValue
-
-runtimeValueMatchesDataTypeName :: Text -> RuntimeValue -> Bool
-runtimeValueMatchesDataTypeName typeName runtimeValue =
-  case runtimeValue of
-    VConstructor valueTypeName _ _ constructorArguments capturedArgs ->
-      identifierText valueTypeName == typeName
-        && constructorIsSaturated constructorArguments capturedArgs
-    _ -> False
-
-runtimeValueMatchesDataTypeApplication :: Name -> [SignatureType] -> RuntimeValue -> Bool
-runtimeValueMatchesDataTypeApplication typeName typeArguments runtimeValue =
-  case runtimeValue of
-    VConstructor valueTypeName typeParameters _ constructorArguments capturedArgs
-      | valueTypeName == typeName,
-        length typeParameters == length typeArguments,
-        constructorIsSaturated constructorArguments capturedArgs ->
-          let typeParameterBindings = Map.fromList (zip (map identifierText typeParameters) typeArguments)
-           in and
-                ( zipWith
-                    (runtimeValueMatchesConstructorArgument typeParameterBindings)
-                    constructorArguments
-                    capturedArgs
-                )
-    _ -> False
-
-runtimeValueExactlyMatchesDataTypeName :: Name -> RuntimeValue -> Bool
-runtimeValueExactlyMatchesDataTypeName typeName runtimeValue =
-  case runtimeValue of
-    VConstructor valueTypeName _ _ constructorArguments capturedArgs ->
-      valueTypeName == typeName
-        && constructorIsSaturated constructorArguments capturedArgs
-    _ -> False
-
-runtimeValueExactlyMatchesDataTypeApplication :: Name -> [SignatureType] -> RuntimeValue -> Bool
-runtimeValueExactlyMatchesDataTypeApplication typeName typeArguments runtimeValue =
-  case runtimeValue of
-    VConstructor valueTypeName typeParameters _ constructorArguments capturedArgs
-      | valueTypeName == typeName,
-        length typeParameters == length typeArguments,
-        constructorIsSaturated constructorArguments capturedArgs ->
-          let typeParameterBindings = Map.fromList (zip (map identifierText typeParameters) typeArguments)
-           in and
-                ( zipWith
-                    (runtimeValueExactlyMatchesConstructorArgument typeParameterBindings)
-                    constructorArguments
-                    capturedArgs
-                )
-    _ -> False
-
-runtimeValueMatchesConstructorArgument :: Map Text SignatureType -> DataConstructorArgument -> RuntimeValue -> Bool
-runtimeValueMatchesConstructorArgument typeParameterBindings constructorArgument runtimeValue =
-  case constructorArgument of
-    DataConstructorArgumentName argumentName ->
-      case constructorArgumentRuntimeHint typeParameterBindings argumentName of
-        Just concreteArgumentType ->
-          runtimeValueMatchesConstraint concreteArgumentType runtimeValue
-        Nothing ->
-          True
-    DataConstructorArgumentOpaque ->
-      True
-
-runtimeValueExactlyMatchesConstructorArgument :: Map Text SignatureType -> DataConstructorArgument -> RuntimeValue -> Bool
-runtimeValueExactlyMatchesConstructorArgument typeParameterBindings constructorArgument runtimeValue =
-  case constructorArgument of
-    DataConstructorArgumentName argumentName ->
-      case Map.lookup (identifierText argumentName) typeParameterBindings of
-        Just concreteArgumentType ->
-          runtimeValueExactlyMatchesConstraint concreteArgumentType runtimeValue
-        Nothing ->
-          True
-    DataConstructorArgumentOpaque ->
-      True
-
-runtimeIntMatchesIntAlias :: RuntimeValue -> Bool
-runtimeIntMatchesIntAlias runtimeValue =
-  case runtimeValue of
-    VInt _ metadata ->
-      case runtimeIntTargetType metadata of
-        Just NumericInt64 -> True
-        Just _ -> False
-        Nothing -> True
-    _ -> False
-
-runtimeIntMatchesTarget :: NumericType -> RuntimeValue -> Bool
-runtimeIntMatchesTarget targetType runtimeValue =
-  case runtimeValue of
-    VInt integerValue metadata ->
-      case runtimeIntTargetType metadata of
-        Just runtimeTarget -> runtimeTarget == targetType
-        Nothing -> integerValueMatchesTarget targetType integerValue
-    _ -> False
-
-integerValueMatchesTarget :: NumericType -> Integer -> Bool
-integerValueMatchesTarget targetType integerValue =
-  case numericTypeIntegerBounds targetType of
-    Just bounds -> integerValueWithinBounds integerValue bounds
-    Nothing -> False
-
-runtimeFloatMatchesFloatAlias :: RuntimeValue -> Bool
-runtimeFloatMatchesFloatAlias runtimeValue =
-  case runtimeValue of
-    VFloat _ metadata ->
-      case runtimeFloatTargetType metadata of
-        Just NumericFloat64 -> True
-        Just _ -> False
-        Nothing -> True
-    _ -> False
-
-runtimeFloatHasTarget :: NumericType -> RuntimeValue -> Bool
-runtimeFloatHasTarget targetType runtimeValue =
-  case runtimeValue of
-    VFloat _ metadata ->
-      case runtimeFloatTargetType metadata of
-        Just runtimeTarget -> runtimeTarget == targetType
-        Nothing -> targetType == NumericFloat64
-    _ -> False
-
-isRuntimeBool :: RuntimeValue -> Bool
-isRuntimeBool runtimeValue =
-  case runtimeValue of
-    VBool {} -> True
-    _ -> False
-
-isRuntimeChar :: RuntimeValue -> Bool
-isRuntimeChar runtimeValue =
-  case runtimeValue of
-    VChar {} -> True
-    _ -> False
-
-isRuntimeText :: RuntimeValue -> Bool
-isRuntimeText runtimeValue =
-  case runtimeValue of
-    VText {} -> True
-    _ -> False
-
--- | Constructor values are curried like builtins until their declared arity is
--- saturated; extra applications are runtime errors.
-applyConstructor :: Name -> [Name] -> Name -> [DataConstructorArgument] -> [RuntimeValue] -> Either Diagnostic RuntimeValue
-applyConstructor typeName typeParameters constructorName constructorArguments arguments
-  | length arguments <= constructorArity =
-      Right (VConstructor typeName typeParameters constructorName constructorArguments arguments)
-  | otherwise =
-      Left
-        ( runtimeDiagnostic
-            "E3023"
-            ( "runtime constructor '"
-                <> identifierText constructorName
-                <> "' expected "
-                <> renderArityCount constructorArity
-                <> " but received "
-                <> renderArityCount (length arguments)
-            )
-        )
-  where
-    constructorArity = length constructorArguments
-
-renderArityCount :: Int -> Text
-renderArityCount count =
-  Text.pack (show count) <> " " <> argumentWord
-  where
-    argumentWord =
-      if count == 1
-        then "argument"
-        else "arguments"
-
--- | Evaluate builtin semantics once enough arguments have been collected.
-evalBuiltin :: BuiltinResolutionMode -> Map BindingRuntimeHintKey SignatureType -> BuiltinSymbol -> [RuntimeValue] -> Either Diagnostic RuntimeValue
-evalBuiltin builtinMode bindingTypeHints builtinFunction arguments =
-  case (builtinFunction, arguments) of
-    (_, [value])
-      | Just targetType <- builtinSymbolNumericConversionTarget builtinFunction ->
-          evalNumericConversion builtinFunction targetType value
-    (BuiltinHd, [VList [] _]) ->
-      Left (runtimeDiagnostic "E3009" "runtime primitive 'hd' failed: empty list")
-    (BuiltinHd, [VList (headValue : _) maybeTypeHint]) ->
-      case maybeTypeHint of
-        Just (TypeList elementType) ->
-          applyRuntimeTypeHint elementType headValue
-        _ ->
-          Right headValue
-    (BuiltinHd, [other]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3011"
-            ("runtime primitive 'hd' expects a list argument, found " <> renderRuntimeType other)
-        )
-    (BuiltinTl, [VList [] _]) ->
-      Left (runtimeDiagnostic "E3010" "runtime primitive 'tl' failed: empty list")
-    (BuiltinTl, [VList (_ : tailValues) maybeTypeHint]) ->
-      Right (VList tailValues maybeTypeHint)
-    (BuiltinTl, [other]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3012"
-            ("runtime primitive 'tl' expects a list argument, found " <> renderRuntimeType other)
-        )
-    (BuiltinMap, [mapper, collection])
-      | not (isFunctionValue mapper) ->
-          Left
-            ( runtimeDiagnostic
-                "E3015"
-                ("runtime primitive 'map' expects a function as its first argument, found " <> renderRuntimeType mapper)
-            )
-      | otherwise ->
-          case collection of
-            VList elements maybeCollectionTypeHint -> do
-              mappedElements <- mapM (applyRuntimeFunction builtinMode bindingTypeHints mapper) elements
-              let maybeMappedTypeHint = TypeList <$> runtimeMapResultElementType mapper maybeCollectionTypeHint
-              Right (VList mappedElements maybeMappedTypeHint)
-            other ->
-              Left
-                ( runtimeDiagnostic
-                    "E3013"
-                    ("runtime primitive 'map' expects a list as its second argument, found " <> renderRuntimeType other)
-                )
-    (BuiltinFilter, [predicate, collection])
-      | not (isFunctionValue predicate) ->
-          Left
-            ( runtimeDiagnostic
-                "E3017"
-                ("runtime primitive 'filter' expects a function as its first argument, found " <> renderRuntimeType predicate)
-            )
-      | otherwise ->
-          case collection of
-            VList elements maybeTypeHint ->
-              (`VList` maybeTypeHint) <$> filterElements builtinMode bindingTypeHints predicate elements
-            other ->
-              Left
-                ( runtimeDiagnostic
-                    "E3018"
-                    ("runtime primitive 'filter' expects a list as its second argument, found " <> renderRuntimeType other)
-                )
-    -- Stub-v1 keeps `print!` side effects out of runtime plumbing; it returns
-    -- its evaluated argument so expression pipelines remain deterministic.
-    (BuiltinPrint, [value]) ->
-      Right value
-    (BuiltinListPrependRaw, [value, VList elements maybeTypeHint]) ->
-      Right (VList (value : elements) maybeTypeHint)
-    (BuiltinListPrependRaw, [_, other]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3032"
-            ("runtime primitive 'listPrependRaw' expects a list as its second argument, found " <> renderRuntimeType other)
-        )
-    (BuiltinCharToUInt32, [VChar value]) ->
-      Right (VInt (fromIntegral (ord value)) (targetedIntMetadata NumericUInt32))
-    (BuiltinCharToUInt32, [other]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3033"
-            ("runtime primitive 'charToUInt32' expects a Char argument, found " <> renderRuntimeType other)
-        )
-    (BuiltinCharFromUInt32Raw, [value@(VInt scalar _)])
-      | runtimeIntMatchesTarget NumericUInt32 value ->
-          let listTypeHint = Just (TypeList TypeChar)
-           in
-            if scalar <= 0x10FFFF && not (scalar >= 0xD800 && scalar <= 0xDFFF)
-              then Right (VList [VChar (chr (fromInteger scalar))] listTypeHint)
-              else Right (VList [] listTypeHint)
-    (BuiltinCharFromUInt32Raw, [other]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3034"
-            ("runtime primitive 'charFromUInt32Raw' expects a UInt32 argument, found " <> renderRuntimeType other)
-        )
-    (BuiltinCharIsAlpha, [VChar value]) -> Right (VBool (isAlpha value))
-    (BuiltinCharIsAlphaNum, [VChar value]) -> Right (VBool (isAlphaNum value))
-    (BuiltinCharIsDigit, [VChar value]) -> Right (VBool (isDigit value))
-    (BuiltinCharIsSpace, [VChar value]) -> Right (VBool (isSpace value))
-    (BuiltinCharIsHexDigit, [VChar value]) -> Right (VBool (isHexDigit value))
-    (builtin@BuiltinCharIsAlpha, [other]) -> invalidCharPredicate builtin other
-    (builtin@BuiltinCharIsAlphaNum, [other]) -> invalidCharPredicate builtin other
-    (builtin@BuiltinCharIsDigit, [other]) -> invalidCharPredicate builtin other
-    (builtin@BuiltinCharIsSpace, [other]) -> invalidCharPredicate builtin other
-    (builtin@BuiltinCharIsHexDigit, [other]) -> invalidCharPredicate builtin other
-    (BuiltinTextLength, [VText textValue]) ->
-      Right (VInt (fromIntegral (Text.length textValue)) untypedIntMetadata)
-    (BuiltinTextLength, [other]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3028"
-            ("runtime primitive 'textLength' expects a Text argument, found " <> renderRuntimeType other)
-        )
-    (BuiltinTextUnconsRaw, [VText textValue]) ->
-      let listTypeHint = Just (TypeList (TypeTuple [TypeChar, TypeText]))
-       in
-        case Text.uncons textValue of
-          Nothing ->
-            Right (VList [] listTypeHint)
-          Just (first, rest) ->
-            Right (VList [VTuple [VChar first, VText rest]] listTypeHint)
-    (BuiltinTextUnconsRaw, [other]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3029"
-            ("runtime primitive 'textUnconsRaw' expects a Text argument, found " <> renderRuntimeType other)
-        )
-    (BuiltinTextAppend, [VText left, VText right]) ->
-      Right (VText (left <> right))
-    (BuiltinTextAppend, [left, right]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3036"
-            ( "runtime primitive 'textAppend' expects Text arguments, found "
-                <> renderRuntimeType left
-                <> " and "
-                <> renderRuntimeType right
-            )
-        )
-    (BuiltinTextAppendChar, [VText textValue, VChar charValue]) ->
-      Right (VText (Text.snoc textValue charValue))
-    (BuiltinTextAppendChar, [textValue, charValue]) ->
-      Left
-        ( runtimeDiagnostic
-            "E3037"
-            ( "runtime primitive 'textAppendChar' expects Text then Char, found "
-                <> renderRuntimeType textValue
-                <> " and "
-                <> renderRuntimeType charValue
-            )
-        )
-    _ ->
-      Left
-        ( runtimeDiagnostic
-            "E3016"
-            ("runtime primitive '" <> builtinSymbolName builtinFunction <> "' received invalid arguments")
-        )
-
-invalidCharPredicate :: BuiltinSymbol -> RuntimeValue -> Either Diagnostic RuntimeValue
-invalidCharPredicate builtin other =
-  Left
-    ( runtimeDiagnostic
-        "E3035"
-        ( "runtime primitive '"
-            <> builtinSymbolName builtin
-            <> "' expects a Char argument, found "
-            <> renderRuntimeType other
-        )
-    )
-
-evalNumericConversion :: BuiltinSymbol -> NumericType -> RuntimeValue -> Either Diagnostic RuntimeValue
-evalNumericConversion builtinFunction targetType value =
-  case value of
-    VExplicitTypeApplication _ innerValue ->
-      evalNumericConversion builtinFunction targetType innerValue
-    VExplicitResultHints _ innerValue ->
-      evalNumericConversion builtinFunction targetType innerValue
-    VTyped _ innerValue ->
-      evalNumericConversion builtinFunction targetType innerValue
-    VInt integerValue _ ->
-      convertIntegerToNumericTarget builtinFunction targetType integerValue
-    VFloat floatValue floatMetadata ->
-      convertFloatToNumericTarget builtinFunction targetType floatValue (runtimeFloatLiteralSource floatMetadata)
-    other ->
-      Left
-        ( runtimeDiagnostic
-            "E3024"
-            ( "runtime numeric conversion '"
-                <> builtinSymbolName builtinFunction
-                <> "' expects a numeric value, found "
-                <> renderRuntimeType other
-            )
-        )
-
-numericConversionBuiltinForTarget :: NumericType -> BuiltinSymbol
-numericConversionBuiltinForTarget targetType =
-  case targetType of
-    NumericInt8 -> BuiltinToInt8
-    NumericInt16 -> BuiltinToInt16
-    NumericInt32 -> BuiltinToInt32
-    NumericInt64 -> BuiltinToInt64
-    NumericUInt8 -> BuiltinToUInt8
-    NumericUInt16 -> BuiltinToUInt16
-    NumericUInt32 -> BuiltinToUInt32
-    NumericUInt64 -> BuiltinToUInt64
-    NumericFloat16 -> BuiltinToFloat16
-    NumericFloat32 -> BuiltinToFloat32
-    NumericFloat64 -> BuiltinToFloat64
-
-convertIntegerToNumericTarget :: BuiltinSymbol -> NumericType -> Integer -> Either Diagnostic RuntimeValue
-convertIntegerToNumericTarget builtinFunction targetType integerValue =
-  case numericTypeIntegerBounds targetType of
-    Just bounds ->
-      if integerValueWithinBounds integerValue bounds
-        then Right (VInt integerValue (targetedIntMetadata targetType))
-        else Left (numericConversionRangeDiagnostic builtinFunction targetType integerValue bounds)
-    Nothing ->
-      convertIntegerToFloatTarget builtinFunction targetType integerValue
-
-convertFloatToNumericTarget :: BuiltinSymbol -> NumericType -> Double -> Maybe FractionalLiteralSource -> Either Diagnostic RuntimeValue
-convertFloatToNumericTarget builtinFunction targetType floatValue literalSource
-  | isNaN floatValue || isInfinite floatValue =
-      Left
-        ( runtimeDiagnostic
-            "E3024"
-            ( "runtime numeric conversion '"
-                <> builtinSymbolName builtinFunction
-                <> "' cannot convert non-finite Float value"
-            )
-        )
-  | otherwise =
-      case numericTypeIntegerBounds targetType of
-        Just bounds ->
-          convertFloatToIntegerTarget builtinFunction targetType floatValue literalSource bounds
-        Nothing ->
-          convertFiniteFloatToFloatTarget builtinFunction targetType floatValue literalSource
-
-convertFloatToIntegerTarget ::
-  BuiltinSymbol ->
-  NumericType ->
-  Double ->
-  Maybe FractionalLiteralSource ->
-  (Integer, Integer) ->
-  Either Diagnostic RuntimeValue
-convertFloatToIntegerTarget builtinFunction targetType floatValue literalSource bounds =
-  case literalSource of
-    Just source ->
-      case fractionalLiteralIntegralValue source of
-        Just integralValue
-          | integerValueWithinBounds integralValue bounds ->
-              Right (VInt integralValue (targetedIntMetadata targetType))
-        _ ->
-          Left (numericConversionFloatToIntegralDiagnostic builtinFunction targetType floatValue bounds)
-    Nothing ->
-      -- `round` is half-to-even, but the equality check below rejects every
-      -- non-integral value instead of observing a rounding mode.
-      let roundedInteger = round floatValue :: Integer
-       in
-        if fromInteger roundedInteger == floatValue && integerValueWithinBounds roundedInteger bounds
-          then Right (VInt roundedInteger (targetedIntMetadata targetType))
-          else Left (numericConversionFloatToIntegralDiagnostic builtinFunction targetType floatValue bounds)
-
-convertIntegerToFloatTarget :: BuiltinSymbol -> NumericType -> Integer -> Either Diagnostic RuntimeValue
-convertIntegerToFloatTarget builtinFunction targetType integerValue =
-  if integerExceedsFloatTarget targetType integerValue
-    then Left (numericConversionFloatOverflowDiagnostic builtinFunction targetType)
-    else
-      let floatValue = fromInteger integerValue :: Double
-       in
-        if isInfinite floatValue || exceedsFloatTarget targetType floatValue
-          then Left (numericConversionFloatOverflowDiagnostic builtinFunction targetType)
-          else Right (VFloat (roundFloatTarget targetType floatValue) (targetedFloatMetadata targetType))
-
-integerExceedsFloatTarget :: NumericType -> Integer -> Bool
-integerExceedsFloatTarget targetType integerValue =
-  case numericTypeFloatMax targetType of
-    Just maxMagnitude ->
-      abs integerValue > (floor maxMagnitude :: Integer)
-    Nothing -> False
-
-convertFiniteFloatToFloatTarget :: BuiltinSymbol -> NumericType -> Double -> Maybe FractionalLiteralSource -> Either Diagnostic RuntimeValue
-convertFiniteFloatToFloatTarget builtinFunction targetType floatValue literalSource =
-  if exceedsFloatTarget targetType floatValue || sourceExceedsFloatTarget targetType literalSource
-    then Left (numericConversionFloatOverflowDiagnostic builtinFunction targetType)
-    else Right (VFloat (roundFloatTarget targetType floatValue) (targetedFloatMetadataWithSource targetType literalSource))
-
-roundFloatTarget :: NumericType -> Double -> Double
-roundFloatTarget targetType value =
-  case targetType of
-    NumericFloat16 -> roundFloat16 value
-    NumericFloat32 -> realToFrac (realToFrac value :: Float)
-    _ -> value
-
-roundFloat16 :: Double -> Double
-roundFloat16 value
-  | value == 0 = 0
-  | magnitude < (halfMinSubnormal / 2.0) = 0
-  | magnitude < halfMinNormal =
-      withSign (fromInteger (round (magnitude / halfMinSubnormal) :: Integer) * halfMinSubnormal)
-  | otherwise =
-      let exponentValue = floor (logBase 2 magnitude) :: Int
-          unit = 2.0 ** fromIntegral (exponentValue - 10)
-          roundedMagnitude = fromInteger (round (magnitude / unit) :: Integer) * unit
-       in withSign (min float16MaxFinite roundedMagnitude)
-  where
-    magnitude = abs value
-    float16MaxFinite = 65504.0 :: Double
-    halfMinNormal = 2.0 ** (-14.0 :: Double)
-    halfMinSubnormal = 2.0 ** (-24.0 :: Double)
-    withSign roundedMagnitude =
-      if value < 0
-        then negate roundedMagnitude
-        else roundedMagnitude
-
-exceedsFloatTarget :: NumericType -> Double -> Bool
-exceedsFloatTarget targetType value =
-  case numericTypeFloatMax targetType of
-    Just maxMagnitude -> abs value > maxMagnitude
-    Nothing -> False
-
-sourceExceedsFloatTarget :: NumericType -> Maybe FractionalLiteralSource -> Bool
-sourceExceedsFloatTarget targetType literalSource =
-  case (numericTypeFloatMax targetType, literalSource) of
-    (Just maxMagnitude, Just source) ->
-      fractionalLiteralExceedsMagnitude source maxMagnitude
-    _ -> False
-
-integerValueWithinBounds :: Integer -> (Integer, Integer) -> Bool
-integerValueWithinBounds value (lowerBound, upperBound) =
-  value >= lowerBound && value <= upperBound
-
-numericConversionRangeDiagnostic :: BuiltinSymbol -> NumericType -> Integer -> (Integer, Integer) -> Diagnostic
-numericConversionRangeDiagnostic builtinFunction targetType value (lowerBound, upperBound) =
-  runtimeDiagnostic
-    "E3024"
-    ( "runtime numeric conversion '"
-        <> builtinSymbolName builtinFunction
-        <> "' failed: integer value "
-        <> Text.pack (show value)
-        <> " outside "
-        <> renderNumericTypeName targetType
-        <> " range "
-        <> Text.pack (show lowerBound)
-        <> ".."
-        <> Text.pack (show upperBound)
-    )
-
-numericConversionFloatToIntegralDiagnostic :: BuiltinSymbol -> NumericType -> Double -> (Integer, Integer) -> Diagnostic
-numericConversionFloatToIntegralDiagnostic builtinFunction targetType value (lowerBound, upperBound) =
-  runtimeDiagnostic
-    "E3024"
-    ( "runtime numeric conversion '"
-        <> builtinSymbolName builtinFunction
-        <> "' failed: Float value "
-        <> Text.pack (show value)
-        <> " must be integral and inside "
-        <> renderNumericTypeName targetType
-        <> " range "
-        <> Text.pack (show lowerBound)
-        <> ".."
-        <> Text.pack (show upperBound)
-    )
-
-numericConversionFloatOverflowDiagnostic :: BuiltinSymbol -> NumericType -> Diagnostic
-numericConversionFloatOverflowDiagnostic builtinFunction targetType =
-  runtimeDiagnostic
-    "E3024"
-    ( "runtime numeric conversion '"
-        <> builtinSymbolName builtinFunction
-        <> "' failed: value cannot be represented as finite "
-        <> renderNumericTypeName targetType
-    )
-
--- | Evaluate filter predicates element-by-element and enforce that each
--- predicate application returns a Bool.
-filterElements :: BuiltinResolutionMode -> Map BindingRuntimeHintKey SignatureType -> RuntimeValue -> [RuntimeValue] -> Either Diagnostic [RuntimeValue]
-filterElements builtinMode bindingTypeHints predicate values = do
-  results <- mapM applyPredicate values
-  pure [value | (value, True) <- results]
-  where
-    -- Preserve runtime safety for partially-known function values that can slip
-    -- past compile-time checks in direct `evaluateRuntimeExpr` tests.
-    applyPredicate :: RuntimeValue -> Either Diagnostic (RuntimeValue, Bool)
-    applyPredicate value = do
-      predicateResult <- applyRuntimeFunction builtinMode bindingTypeHints predicate value
-      case predicateResult of
-        VBool shouldKeep -> Right (value, shouldKeep)
-        other ->
-          Left
-            ( runtimeDiagnostic
-                "E3019"
-                ("runtime primitive 'filter' predicate must return Bool, found " <> renderRuntimeType other)
-            )
-
-runtimeFunctionResultType :: RuntimeValue -> Maybe SignatureType
-runtimeFunctionResultType runtimeValue =
-  case runtimeValue of
-    VExplicitTypeApplication _ innerValue ->
-      runtimeFunctionResultType innerValue
-    VExplicitResultHints _ innerValue ->
-      runtimeFunctionResultType innerValue
-    VTyped (TypeFunction _ resultType) _ ->
-      Just resultType
-    VClosure _ _ _ _ (Just (TypeFunction _ resultType)) _ ->
-      Just resultType
-    _ ->
-      Nothing
-
-runtimeMapResultElementType :: RuntimeValue -> Maybe SignatureType -> Maybe SignatureType
-runtimeMapResultElementType mapper maybeCollectionTypeHint =
-  case runtimeFunctionResultType mapper of
-    Just resultType ->
-      Just resultType
-    Nothing ->
-      runtimeBuiltinMapResultElementType mapper maybeCollectionTypeHint
-
-runtimeBuiltinMapResultElementType :: RuntimeValue -> Maybe SignatureType -> Maybe SignatureType
-runtimeBuiltinMapResultElementType mapper maybeCollectionTypeHint =
-  case (mapper, maybeCollectionTypeHint) of
-    (VBuiltin BuiltinHd [], Just (TypeList (TypeList elementType))) ->
-      Just elementType
-    (VClosure _ _ parameterName (EVar resultName) Nothing _, Just (TypeList elementType))
-      | resultName == parameterName ->
-          Just elementType
-    _ ->
-      Nothing
-
-attachDefaultBindingIntegerTarget :: RuntimeValue -> Either Diagnostic RuntimeValue
-attachDefaultBindingIntegerTarget runtimeValue =
-  case runtimeValue of
-    VInt integerValue metadata
-      | runtimeIntTargetType metadata == Nothing,
-        integerValueMatchesTarget NumericInt64 integerValue ->
-          Right (VInt integerValue (targetedIntMetadata NumericInt64))
-    VList elements maybeTypeHint ->
-      (`VList` maybeTypeHint) <$> traverse attachDefaultBindingIntegerTarget elements
-    VTuple elements ->
-      VTuple <$> traverse attachDefaultBindingIntegerTarget elements
-    VBuiltin builtinSymbol capturedArgs ->
-      VBuiltin builtinSymbol <$> traverse attachDefaultBindingIntegerTarget capturedArgs
-    VOperator operatorSymbol capturedArgs ->
-      VOperator operatorSymbol <$> traverse attachDefaultBindingIntegerTarget capturedArgs
-    VSectionLeft operatorSymbol operand ->
-      VSectionLeft operatorSymbol <$> attachDefaultBindingIntegerTarget operand
-    VSectionRight operatorSymbol operand ->
-      VSectionRight operatorSymbol <$> attachDefaultBindingIntegerTarget operand
-    VConstructor typeName typeParameters constructorName constructorArguments capturedArgs ->
-      VConstructor typeName typeParameters constructorName constructorArguments
-        <$> traverse attachDefaultBindingIntegerTarget capturedArgs
-    VQualifiedMethod methodKey classParameter methodSignature candidates capturedArgs ->
-      VQualifiedMethod methodKey classParameter methodSignature candidates
-        <$> traverse attachDefaultBindingIntegerTarget capturedArgs
-    VTyped typeHint innerValue
-      | TypeFunction {} <- typeHint ->
-          Right (VTyped typeHint innerValue)
-      | otherwise ->
-          VTyped typeHint <$> attachDefaultBindingIntegerTarget innerValue
-    VExplicitTypeApplication typeHint innerValue ->
-      VExplicitTypeApplication typeHint <$> attachDefaultBindingIntegerTarget innerValue
-    VExplicitResultHints hints innerValue ->
-      attachRuntimeExplicitResultHints hints <$> attachDefaultBindingIntegerTarget innerValue
-    _ ->
-      Right runtimeValue
-
-isFunctionValue :: RuntimeValue -> Bool
-isFunctionValue value =
-  case value of
-    VExplicitTypeApplication _ innerValue -> isFunctionValue innerValue
-    VExplicitResultHints _ innerValue -> isFunctionValue innerValue
-    VTyped _ innerValue -> isFunctionValue innerValue
-    VSectionLeft {} -> True
-    VSectionRight {} -> True
-    VClosure {} -> True
-    VBuiltin {} -> True
-    VOperator {} -> True
-    VConstructor _ _ _ constructorArguments capturedArgs ->
-      not (constructorIsSaturated constructorArguments capturedArgs)
-    VQualifiedMethod {} -> True
-    _ -> False
-
--- | Evaluate the builtin operator subset supported by the runtime.
-evalBinary :: BuiltinResolutionMode -> Map BindingRuntimeHintKey SignatureType -> Text -> RuntimeValue -> RuntimeValue -> Either Diagnostic RuntimeValue
-evalBinary builtinMode bindingTypeHints operatorSymbol leftValue rightValue
-  | isStrictEqualityOperator operatorSymbol,
-    isFunctionValue leftValue || isFunctionValue rightValue =
-      Left (runtimeCallableEqualityDiagnostic operatorSymbol leftValue rightValue)
-  | otherwise =
-  case (operatorSymbol, leftValue, rightValue) of
-    ("$", functionValue, argumentValue) ->
-      applyRuntimeFunction builtinMode bindingTypeHints functionValue argumentValue
-    (_, VTyped leftTypeHint leftInnerValue, _)
-      | isStrictEqualityOperator operatorSymbol,
-        runtimeTypeHintRequiresStructuralEquality leftTypeHint ->
-          evalStructuralEquality operatorSymbol leftValue rightValue
-      | otherwise ->
-          preserveLeftTypedNumericOperatorResult operatorSymbol leftTypeHint
-            =<< evalBinary builtinMode bindingTypeHints operatorSymbol leftInnerValue rightValue
-    (_, _, VTyped rightTypeHint rightInnerValue)
-      | isStrictEqualityOperator operatorSymbol,
-        runtimeTypeHintRequiresStructuralEquality rightTypeHint ->
-          evalStructuralEquality operatorSymbol leftValue rightValue
-      | otherwise ->
-          preserveRightTypedNumericOperatorResult operatorSymbol leftValue rightTypeHint
-            =<< evalBinary builtinMode bindingTypeHints operatorSymbol leftValue rightInnerValue
-    ("+", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerArithmetic "+" leftMetadata rightMetadata (leftInt + rightInt)
-    ("-", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerArithmetic "-" leftMetadata rightMetadata (leftInt - rightInt)
-    ("*", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerArithmetic "*" leftMetadata rightMetadata (leftInt * rightInt)
-    ("/", VInt _ _, VInt 0 _) ->
-      Left (runtimeDiagnostic "E3001" "runtime primitive '/' failed: division by zero")
-    ("/", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerArithmetic "/" leftMetadata rightMetadata (leftInt `div` rightInt)
-    ("+", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatArithmetic "+" leftMetadata rightMetadata (leftFloat + rightFloat)
-    ("-", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatArithmetic "-" leftMetadata rightMetadata (leftFloat - rightFloat)
-    ("*", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatArithmetic "*" leftMetadata rightMetadata (leftFloat * rightFloat)
-    ("/", VFloat _ _, VFloat rightFloat _)
-      | floatIsZero rightFloat ->
-          Left (runtimeDiagnostic "E3001" "runtime primitive '/' failed: division by zero")
-    ("/", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatArithmetic "/" leftMetadata rightMetadata (leftFloat / rightFloat)
-    ("+", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64PromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Arithmetic "+" rightMetadata leftInt rightFloat (+)
-    ("+", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64PromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerArithmetic "+" leftMetadata leftFloat rightInt (+)
-    ("-", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64PromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Arithmetic "-" rightMetadata leftInt rightFloat (-)
-    ("-", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64PromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerArithmetic "-" leftMetadata leftFloat rightInt (-)
-    ("*", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64PromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Arithmetic "*" rightMetadata leftInt rightFloat (*)
-    ("*", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64PromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerArithmetic "*" leftMetadata leftFloat rightInt (*)
-    ("/", VInt _ leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64PromotionAccepted leftMetadata rightMetadata,
-        floatIsZero rightFloat ->
-          Left (runtimeDiagnostic "E3001" "runtime primitive '/' failed: division by zero")
-    ("/", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64PromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Arithmetic "/" rightMetadata leftInt rightFloat (/)
-    ("/", VFloat _ leftMetadata, VInt 0 rightMetadata)
-      | runtimeIntFloat64PromotionAccepted rightMetadata leftMetadata ->
-          Left (runtimeDiagnostic "E3001" "runtime primitive '/' failed: division by zero")
-    ("/", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64PromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerArithmetic "/" leftMetadata leftFloat rightInt (/)
-    ("<", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerPredicate "<" leftInt leftMetadata rightInt rightMetadata (leftInt < rightInt)
-    ("<=", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerPredicate "<=" leftInt leftMetadata rightInt rightMetadata (leftInt <= rightInt)
-    (">", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerPredicate ">" leftInt leftMetadata rightInt rightMetadata (leftInt > rightInt)
-    (">=", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerPredicate ">=" leftInt leftMetadata rightInt rightMetadata (leftInt >= rightInt)
-    ("<", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatPredicate "<" leftMetadata rightMetadata (leftFloat < rightFloat)
-    ("<=", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatPredicate "<=" leftMetadata rightMetadata (leftFloat <= rightFloat)
-    (">", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatPredicate ">" leftMetadata rightMetadata (leftFloat > rightFloat)
-    (">=", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatPredicate ">=" leftMetadata rightMetadata (leftFloat >= rightFloat)
-    ("<", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Predicate "<" leftInt rightFloat (<)
-    ("<", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerPredicate "<" leftFloat rightInt (<)
-    ("<=", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Predicate "<=" leftInt rightFloat (<=)
-    ("<=", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerPredicate "<=" leftFloat rightInt (<=)
-    (">", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Predicate ">" leftInt rightFloat (>)
-    (">", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerPredicate ">" leftFloat rightInt (>)
-    (">=", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Predicate ">=" leftInt rightFloat (>=)
-    (">=", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerPredicate ">=" leftFloat rightInt (>=)
-    ("==", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerEquality "==" leftInt leftMetadata rightInt rightMetadata
-    ("==", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatPredicate "==" leftMetadata rightMetadata (leftFloat == rightFloat)
-    ("==", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Equality "==" leftInt rightFloat
-    ("==", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerEquality "==" leftFloat rightInt
-    ("==", VBool leftBool, VBool rightBool) -> Right (VBool (leftBool == rightBool))
-    ("==", VChar leftChar, VChar rightChar) -> Right (VBool (leftChar == rightChar))
-    ("==", VText leftText, VText rightText) -> Right (VBool (leftText == rightText))
-    ("==", VList {}, VList {}) -> evalStructuralEquality "==" leftValue rightValue
-    ("==", VTuple {}, VTuple {}) -> evalStructuralEquality "==" leftValue rightValue
-    ("==", VConstructor {}, VConstructor {}) -> evalStructuralEquality "==" leftValue rightValue
-    ("!=", VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      evalIntegerEquality "!=" leftInt leftMetadata rightInt rightMetadata
-    ("!=", VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      evalFloatPredicate "!=" leftMetadata rightMetadata (leftFloat /= rightFloat)
-    ("!=", VInt leftInt leftMetadata, VFloat rightFloat rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted leftMetadata rightMetadata ->
-          evalIntegerFloat64Equality "!=" leftInt rightFloat
-    ("!=", VFloat leftFloat leftMetadata, VInt rightInt rightMetadata)
-      | runtimeIntFloat64ComparisonPromotionAccepted rightMetadata leftMetadata ->
-          evalFloat64IntegerEquality "!=" leftFloat rightInt
-    ("!=", VBool leftBool, VBool rightBool) -> Right (VBool (leftBool /= rightBool))
-    ("!=", VChar leftChar, VChar rightChar) -> Right (VBool (leftChar /= rightChar))
-    ("!=", VText leftText, VText rightText) -> Right (VBool (leftText /= rightText))
-    ("!=", VList {}, VList {}) -> evalStructuralEquality "!=" leftValue rightValue
-    ("!=", VTuple {}, VTuple {}) -> evalStructuralEquality "!=" leftValue rightValue
-    ("!=", VConstructor {}, VConstructor {}) -> evalStructuralEquality "!=" leftValue rightValue
-    _ ->
-      Left
-        ( runtimeDiagnostic
-            "E3007"
-            ( "runtime primitive '"
-                <> operatorSymbol
-                <> "' cannot be applied to "
-                <> renderRuntimeType leftValue
-                <> " and "
-                <> renderRuntimeType rightValue
-            )
-        )
-
-isStrictEqualityOperator :: Text -> Bool
-isStrictEqualityOperator operatorSymbol =
-  operatorSymbol == "==" || operatorSymbol == "!="
-
-preserveLeftTypedNumericOperatorResult :: Text -> SignatureType -> RuntimeValue -> Either Diagnostic RuntimeValue
-preserveLeftTypedNumericOperatorResult operatorSymbol typeHint runtimeValue
-  | numericArithmeticOperator operatorSymbol,
-    numericAliasTypeHint typeHint,
-    runtimeValueMatchesConstraint typeHint runtimeValue =
-      applyRuntimeTypeHint typeHint runtimeValue
-  | otherwise =
-      Right runtimeValue
-
-preserveRightTypedNumericOperatorResult :: Text -> RuntimeValue -> SignatureType -> RuntimeValue -> Either Diagnostic RuntimeValue
-preserveRightTypedNumericOperatorResult operatorSymbol leftValue typeHint runtimeValue
-  | numericArithmeticOperator operatorSymbol,
-    numericAliasTypeHint typeHint,
-    not (runtimeValueHasTargetedNumericMetadata leftValue),
-    runtimeValueMatchesConstraint typeHint runtimeValue =
-      applyRuntimeTypeHint typeHint runtimeValue
-  | otherwise =
-      Right runtimeValue
-
-numericArithmeticOperator :: Text -> Bool
-numericArithmeticOperator operatorSymbol =
-  operatorSymbol == "+" || operatorSymbol == "-" || operatorSymbol == "*" || operatorSymbol == "/"
-
-numericAliasTypeHint :: SignatureType -> Bool
-numericAliasTypeHint typeHint =
-  case typeHint of
-    TypeInt -> True
-    TypeFloat -> True
-    TypeName typeName ->
-      identifierText typeName == "Int" || identifierText typeName == "Float"
-    _ ->
-      False
-
-runtimeValueHasTargetedNumericMetadata :: RuntimeValue -> Bool
-runtimeValueHasTargetedNumericMetadata runtimeValue =
-  case runtimeValue of
-    VInt _ metadata ->
-      runtimeIntTargetType metadata /= Nothing
-    VFloat _ metadata ->
-      runtimeFloatTargetType metadata /= Nothing
-    _ ->
-      False
-
-runtimeTypeHintRequiresStructuralEquality :: SignatureType -> Bool
-runtimeTypeHintRequiresStructuralEquality signatureType =
-  case signatureType of
-    TypeApplication {} -> True
-    TypeList {} -> True
-    TypeTuple {} -> True
-    _ -> False
-
-runtimeCallableEqualityDiagnostic :: Text -> RuntimeValue -> RuntimeValue -> Diagnostic
-runtimeCallableEqualityDiagnostic operatorSymbol leftValue rightValue =
-  runtimeDiagnostic
-    "E3007"
-    ( "runtime primitive '"
-        <> operatorSymbol
-        <> "' cannot compare callable values; callable values are not equality-supported, found "
-        <> renderRuntimeType leftValue
-        <> " and "
-        <> renderRuntimeType rightValue
-    )
-
-evalIntegerArithmetic ::
-  Text ->
-  RuntimeIntMetadata ->
-  RuntimeIntMetadata ->
-  Integer ->
-  Either Diagnostic RuntimeValue
-evalIntegerArithmetic operatorSymbol leftMetadata rightMetadata result = do
-  targetType <- selectIntegerBinaryTarget operatorSymbol leftMetadata rightMetadata
-  evalIntegerBinary operatorSymbol targetType result
-
-selectIntegerBinaryTarget :: Text -> RuntimeIntMetadata -> RuntimeIntMetadata -> Either Diagnostic (Maybe NumericType)
-selectIntegerBinaryTarget operatorSymbol leftMetadata rightMetadata =
-  case (runtimeIntTargetType leftMetadata, runtimeIntTargetType rightMetadata) of
-    (Just leftTarget, Just rightTarget)
-      | leftTarget == rightTarget -> Right (Just leftTarget)
-      | otherwise -> Left (mixedIntegerArithmeticDiagnostic operatorSymbol (Just leftTarget) (Just rightTarget))
-    (Just leftTarget, Nothing) -> Right (Just leftTarget)
-    (Nothing, Just rightTarget) -> Right (Just rightTarget)
-    _ -> Right Nothing
-
-evalIntegerBinary :: Text -> Maybe NumericType -> Integer -> Either Diagnostic RuntimeValue
-evalIntegerBinary operatorSymbol maybeTarget result =
-  case maybeTarget of
-    Just targetType ->
-      case numericTypeIntegerBounds targetType of
-        Just bounds
-          | integerValueWithinBounds result bounds ->
-              Right (VInt result (targetedIntMetadata targetType))
-          | otherwise ->
-              Left (runtimeIntegerArithmeticOverflowDiagnostic operatorSymbol targetType result bounds)
-        Nothing ->
-          Right (VInt result (targetedIntMetadata targetType))
-    Nothing ->
-      Right (VInt result untypedIntMetadata)
-
-mixedIntegerArithmeticDiagnostic :: Text -> Maybe NumericType -> Maybe NumericType -> Diagnostic
-mixedIntegerArithmeticDiagnostic operatorSymbol leftTarget rightTarget =
-  runtimeDiagnostic
-    "E3007"
-    ( "runtime primitive '"
-        <> operatorSymbol
-        <> "' cannot mix "
-        <> renderIntegerOperandTarget leftTarget
-        <> " and "
-        <> renderIntegerOperandTarget rightTarget
-    )
-
-renderIntegerOperandTarget :: Maybe NumericType -> Text
-renderIntegerOperandTarget maybeTarget =
-  case maybeTarget of
-    Just targetType -> renderNumericTypeName targetType
-    Nothing -> "Int"
-
-runtimeIntegerArithmeticOverflowDiagnostic :: Text -> NumericType -> Integer -> (Integer, Integer) -> Diagnostic
-runtimeIntegerArithmeticOverflowDiagnostic operatorSymbol targetType result (lowerBound, upperBound) =
-  runtimeDiagnostic
-    "E3025"
-    ( "runtime primitive '"
-        <> operatorSymbol
-        <> "' failed: integer value "
-        <> Text.pack (show result)
-        <> " outside "
-        <> renderNumericTypeName targetType
-        <> " range "
-        <> Text.pack (show lowerBound)
-        <> ".."
-        <> Text.pack (show upperBound)
-    )
-
-floatIsZero :: Double -> Bool
-floatIsZero value =
-  -- Jazz's finite runtime primitive subset treats both signed zeroes as
-  -- division by zero rather than producing infinities.
-  value == 0
-
-evalFloatArithmetic ::
-  Text ->
-  RuntimeFloatMetadata ->
-  RuntimeFloatMetadata ->
-  Double ->
-  Either Diagnostic RuntimeValue
-evalFloatArithmetic operatorSymbol leftMetadata rightMetadata result = do
-  targetType <- selectFloatBinaryTarget operatorSymbol leftMetadata rightMetadata
-  evalFloatBinary operatorSymbol targetType result
-
-selectFloatBinaryTarget :: Text -> RuntimeFloatMetadata -> RuntimeFloatMetadata -> Either Diagnostic (Maybe NumericType)
-selectFloatBinaryTarget operatorSymbol leftMetadata rightMetadata =
-  case (runtimeFloatTargetType leftMetadata, runtimeFloatTargetType rightMetadata) of
-    (Just leftTarget, Just rightTarget)
-      | leftTarget == rightTarget -> Right (Just leftTarget)
-      | otherwise -> Left (mixedFloatArithmeticDiagnostic operatorSymbol (Just leftTarget) (Just rightTarget))
-    (Just NumericFloat64, Nothing) -> Right (Just NumericFloat64)
-    (Nothing, Just NumericFloat64) -> Right (Just NumericFloat64)
-    (Just targetType, Nothing) -> Left (mixedFloatArithmeticDiagnostic operatorSymbol (Just targetType) Nothing)
-    (Nothing, Just targetType) -> Left (mixedFloatArithmeticDiagnostic operatorSymbol Nothing (Just targetType))
-    (Nothing, Nothing) -> Right Nothing
-
-evalFloatBinary :: Text -> Maybe NumericType -> Double -> Either Diagnostic RuntimeValue
-evalFloatBinary operatorSymbol targetType result
-  | isNaN result || isInfinite result =
-      Left
-        ( runtimeDiagnostic
-            "E3025"
-            ("runtime primitive '" <> operatorSymbol <> "' failed: non-finite Float result")
-        )
-  | Just floatTarget <- targetType,
-    exceedsFloatTarget floatTarget result =
-      Left (runtimeFloatArithmeticOverflowDiagnostic operatorSymbol floatTarget)
-  | Just floatTarget <- targetType =
-      Right (VFloat (roundFloatTarget floatTarget result) (targetedFloatMetadata floatTarget))
-  | otherwise = Right (VFloat result (untypedFloatMetadata Nothing))
-
-runtimeIntFloat64PromotionAccepted :: RuntimeIntMetadata -> RuntimeFloatMetadata -> Bool
-runtimeIntFloat64PromotionAccepted intMetadata floatMetadata =
-  runtimeIntMetadataIsIntegral intMetadata
-    && runtimeFloatMetadataIsFloat64Domain floatMetadata
-
-runtimeIntFloat64ComparisonPromotionAccepted :: RuntimeIntMetadata -> RuntimeFloatMetadata -> Bool
-runtimeIntFloat64ComparisonPromotionAccepted intMetadata floatMetadata =
-  runtimeIntMetadataIsIntegral intMetadata
-    && runtimeFloatMetadataIsFloat64Domain floatMetadata
-
-runtimeIntMetadataIsIntegral :: RuntimeIntMetadata -> Bool
-runtimeIntMetadataIsIntegral intMetadata =
-  case runtimeIntTargetType intMetadata of
-    Just numericType -> numericTypeIsIntegral numericType
-    Nothing -> True
-
-runtimeFloatMetadataIsFloat64Domain :: RuntimeFloatMetadata -> Bool
-runtimeFloatMetadataIsFloat64Domain floatMetadata =
-  case runtimeFloatTargetType floatMetadata of
-    Just NumericFloat64 -> True
-    Nothing -> True
-    Just _ -> False
-
-evalIntegerFloat64Arithmetic :: Text -> RuntimeFloatMetadata -> Integer -> Double -> (Double -> Double -> Double) -> Either Diagnostic RuntimeValue
-evalIntegerFloat64Arithmetic operatorSymbol floatMetadata integerValue floatValue combine = do
-  integerFloat <- promotedIntegerFloat64Operand integerValue
-  evalFloatBinary operatorSymbol (runtimeFloatTargetType floatMetadata) (combine integerFloat floatValue)
-
-evalFloat64IntegerArithmetic :: Text -> RuntimeFloatMetadata -> Double -> Integer -> (Double -> Double -> Double) -> Either Diagnostic RuntimeValue
-evalFloat64IntegerArithmetic operatorSymbol floatMetadata floatValue integerValue combine = do
-  integerFloat <- promotedIntegerFloat64Operand integerValue
-  evalFloatBinary operatorSymbol (runtimeFloatTargetType floatMetadata) (combine floatValue integerFloat)
-
-evalIntegerFloat64Predicate :: Text -> Integer -> Double -> (Double -> Double -> Bool) -> Either Diagnostic RuntimeValue
-evalIntegerFloat64Predicate _ integerValue floatValue predicate = do
-  integerFloat <- promotedIntegerFloat64Operand integerValue
-  pure (VBool (predicate integerFloat floatValue))
-
-evalFloat64IntegerPredicate :: Text -> Double -> Integer -> (Double -> Double -> Bool) -> Either Diagnostic RuntimeValue
-evalFloat64IntegerPredicate _ floatValue integerValue predicate = do
-  integerFloat <- promotedIntegerFloat64Operand integerValue
-  pure (VBool (predicate floatValue integerFloat))
-
-evalIntegerFloat64Equality :: Text -> Integer -> Double -> Either Diagnostic RuntimeValue
-evalIntegerFloat64Equality operatorSymbol integerValue floatValue = do
-  integerFloat <- promotedIntegerFloat64Operand integerValue
-  pure (VBool (float64MixedEqualityResult operatorSymbol integerFloat floatValue))
-
-evalFloat64IntegerEquality :: Text -> Double -> Integer -> Either Diagnostic RuntimeValue
-evalFloat64IntegerEquality operatorSymbol floatValue integerValue = do
-  integerFloat <- promotedIntegerFloat64Operand integerValue
-  pure (VBool (float64MixedEqualityResult operatorSymbol floatValue integerFloat))
-
-float64MixedEqualityResult :: Text -> Double -> Double -> Bool
-float64MixedEqualityResult operatorSymbol leftValue rightValue =
-  if operatorSymbol == "!="
-    then leftValue /= rightValue
-    else leftValue == rightValue
-
-promotedIntegerFloat64Operand :: Integer -> Either Diagnostic Double
-promotedIntegerFloat64Operand integerValue =
-  case convertIntegerToFloatTarget BuiltinToFloat64 NumericFloat64 integerValue of
-    Right (VFloat floatValue _) -> Right floatValue
-    Right _ -> Left (numericConversionFloatOverflowDiagnostic BuiltinToFloat64 NumericFloat64)
-    Left diagnostic -> Left diagnostic
-
-mixedFloatArithmeticDiagnostic :: Text -> Maybe NumericType -> Maybe NumericType -> Diagnostic
-mixedFloatArithmeticDiagnostic operatorSymbol leftTarget rightTarget =
-  runtimeDiagnostic
-    "E3007"
-    ( "runtime primitive '"
-        <> operatorSymbol
-        <> "' cannot mix "
-        <> renderFloatOperandTarget leftTarget
-        <> " and "
-        <> renderFloatOperandTarget rightTarget
-    )
-
-renderFloatOperandTarget :: Maybe NumericType -> Text
-renderFloatOperandTarget maybeTarget =
-  case maybeTarget of
-    Just targetType -> renderNumericTypeName targetType
-    Nothing -> "Float"
-
-runtimeFloatArithmeticOverflowDiagnostic :: Text -> NumericType -> Diagnostic
-runtimeFloatArithmeticOverflowDiagnostic operatorSymbol targetType =
-  runtimeDiagnostic
-    "E3025"
-    ( "runtime primitive '"
-        <> operatorSymbol
-        <> "' failed: value cannot be represented as finite "
-        <> renderNumericTypeName targetType
-    )
-
-evalStructuralEquality :: Text -> RuntimeValue -> RuntimeValue -> Either Diagnostic RuntimeValue
-evalStructuralEquality operatorSymbol leftValue rightValue =
-  if runtimeValueContainsFunction leftValue || runtimeValueContainsFunction rightValue
-    then Left (runtimeCallableEqualityDiagnostic operatorSymbol leftValue rightValue)
-    else
-      case runtimeStructuralEquality leftValue rightValue of
-        Just equalityResult ->
-          Right
-            ( VBool
-                ( if operatorSymbol == "!="
-                    then not equalityResult
-                    else equalityResult
-                )
-            )
-        Nothing ->
-          Left
-            ( runtimeDiagnostic
-                "E3007"
-                ( "runtime primitive '"
-                    <> operatorSymbol
-                    <> "' cannot be applied to "
-                    <> renderRuntimeType leftValue
-                    <> " and "
-                    <> renderRuntimeType rightValue
-                )
-            )
-
-runtimeValueContainsFunction :: RuntimeValue -> Bool
-runtimeValueContainsFunction value =
-  isFunctionValue value
-    || runtimeContainerContainsFunction value
-  where
-    runtimeContainerContainsFunction runtimeValue =
-      case runtimeValue of
-        VList elements _ ->
-          any runtimeValueContainsFunction elements
-        VTuple elements ->
-          any runtimeValueContainsFunction elements
-        VConstructor _ _ _ _ capturedArgs ->
-          any runtimeValueContainsFunction capturedArgs
-        VTyped _ innerValue ->
-          runtimeValueContainsFunction innerValue
-        VExplicitTypeApplication _ innerValue ->
-          runtimeValueContainsFunction innerValue
-        VExplicitResultHints _ innerValue ->
-          runtimeValueContainsFunction innerValue
-        _ ->
-          False
-
-runtimeStructuralEquality :: RuntimeValue -> RuntimeValue -> Maybe Bool
-runtimeStructuralEquality leftValue rightValue =
-  case (leftValue, rightValue) of
-    (VExplicitTypeApplication _ leftInnerValue, _) ->
-      runtimeStructuralEquality leftInnerValue rightValue
-    (_, VExplicitTypeApplication _ rightInnerValue) ->
-      runtimeStructuralEquality leftValue rightInnerValue
-    (VExplicitResultHints _ leftInnerValue, _) ->
-      runtimeStructuralEquality leftInnerValue rightValue
-    (_, VExplicitResultHints _ rightInnerValue) ->
-      runtimeStructuralEquality leftValue rightInnerValue
-    (VTyped leftTypeHint leftInnerValue, VTyped rightTypeHint rightInnerValue)
-      | constraintSignatureTypesCompatible leftTypeHint rightTypeHint ->
-          runtimeStructuralEquality leftInnerValue rightInnerValue
-      | otherwise ->
-          Just False
-    (VTyped _ leftInnerValue, _) ->
-      runtimeStructuralEquality leftInnerValue rightValue
-    (_, VTyped _ rightInnerValue) ->
-      runtimeStructuralEquality leftValue rightInnerValue
-    (VInt leftInt leftMetadata, VInt rightInt rightMetadata) ->
-      runtimeIntegerStructuralEquality leftInt leftMetadata rightInt rightMetadata
-    (VFloat leftFloat leftMetadata, VFloat rightFloat rightMetadata) ->
-      runtimeFloatStructuralEquality leftFloat leftMetadata rightFloat rightMetadata
-    (VBool leftBool, VBool rightBool) -> Just (leftBool == rightBool)
-    (VChar leftChar, VChar rightChar) -> Just (leftChar == rightChar)
-    (VText leftText, VText rightText) -> Just (leftText == rightText)
-    (VList leftElements _, VList rightElements _) ->
-      structuralElementEquality leftElements rightElements
-    (VTuple leftElements, VTuple rightElements) ->
-      structuralElementEquality leftElements rightElements
-    ( VConstructor leftTypeName _ leftName leftConstructorArguments leftArgs,
-      VConstructor rightTypeName _ rightName rightConstructorArguments rightArgs
-      )
-      | constructorIsSaturated leftConstructorArguments leftArgs,
-        constructorIsSaturated rightConstructorArguments rightArgs,
-        leftTypeName == rightTypeName,
-        leftName == rightName,
-        leftConstructorArguments == rightConstructorArguments ->
-          structuralElementEquality leftArgs rightArgs
-      | constructorIsSaturated leftConstructorArguments leftArgs,
-        constructorIsSaturated rightConstructorArguments rightArgs ->
-          Just False
-    _ -> Nothing
-
-structuralElementEquality :: [RuntimeValue] -> [RuntimeValue] -> Maybe Bool
-structuralElementEquality leftElements rightElements
-  | length leftElements /= length rightElements =
-      Just False
-  | otherwise =
-      fmap and
-        (traverse (uncurry runtimeStructuralEquality) (zip leftElements rightElements))
-
-evalIntegerPredicate :: Text -> Integer -> RuntimeIntMetadata -> Integer -> RuntimeIntMetadata -> Bool -> Either Diagnostic RuntimeValue
-evalIntegerPredicate operatorSymbol leftInt leftMetadata rightInt rightMetadata predicateResult =
-  case runtimeIntegerMetadataCompatible leftInt leftMetadata rightInt rightMetadata of
-    True ->
-      Right (VBool predicateResult)
-    False ->
-      Left
-        ( runtimeDiagnostic
-            "E3007"
-            ( "runtime primitive '"
-                <> operatorSymbol
-                <> "' cannot compare "
-                <> renderIntegerOperandTarget (runtimeIntTargetType leftMetadata)
-                <> " and "
-                <> renderIntegerOperandTarget (runtimeIntTargetType rightMetadata)
-            )
-        )
-
-evalIntegerEquality :: Text -> Integer -> RuntimeIntMetadata -> Integer -> RuntimeIntMetadata -> Either Diagnostic RuntimeValue
-evalIntegerEquality operatorSymbol leftInt leftMetadata rightInt rightMetadata =
-  case runtimeIntegerMetadataCompatible leftInt leftMetadata rightInt rightMetadata of
-    True ->
-      Right
-        ( VBool
-            ( if operatorSymbol == "!="
-                then leftInt /= rightInt
-                else leftInt == rightInt
-            )
-        )
-    False ->
-      Left
-        ( runtimeDiagnostic
-            "E3007"
-            ( "runtime primitive '"
-                <> operatorSymbol
-                <> "' cannot compare "
-                <> renderIntegerOperandTarget (runtimeIntTargetType leftMetadata)
-                <> " and "
-                <> renderIntegerOperandTarget (runtimeIntTargetType rightMetadata)
-            )
-        )
-
-evalFloatPredicate :: Text -> RuntimeFloatMetadata -> RuntimeFloatMetadata -> Bool -> Either Diagnostic RuntimeValue
-evalFloatPredicate operatorSymbol leftMetadata rightMetadata predicateResult =
-  if runtimeFloatMetadataCompatible leftMetadata rightMetadata
-    then Right (VBool predicateResult)
-    else
-      Left
-        ( runtimeDiagnostic
-            "E3007"
-            ( "runtime primitive '"
-                <> operatorSymbol
-                <> "' cannot compare "
-                <> renderFloatOperandTarget (runtimeFloatTargetType leftMetadata)
-                <> " and "
-                <> renderFloatOperandTarget (runtimeFloatTargetType rightMetadata)
-            )
-        )
-
-runtimeIntegerStructuralEquality :: Integer -> RuntimeIntMetadata -> Integer -> RuntimeIntMetadata -> Maybe Bool
-runtimeIntegerStructuralEquality leftInt leftMetadata rightInt rightMetadata =
-  if runtimeIntegerMetadataCompatible leftInt leftMetadata rightInt rightMetadata
-    then Just (leftInt == rightInt)
-    else Nothing
-
-runtimeFloatStructuralEquality :: Double -> RuntimeFloatMetadata -> Double -> RuntimeFloatMetadata -> Maybe Bool
-runtimeFloatStructuralEquality leftFloat leftMetadata rightFloat rightMetadata =
-  if runtimeFloatMetadataCompatible leftMetadata rightMetadata
-    then Just (leftFloat == rightFloat)
-    else Nothing
-
-runtimeIntegerMetadataCompatible :: Integer -> RuntimeIntMetadata -> Integer -> RuntimeIntMetadata -> Bool
-runtimeIntegerMetadataCompatible leftInt leftMetadata rightInt rightMetadata =
-  case (runtimeIntTargetType leftMetadata, runtimeIntTargetType rightMetadata) of
-    (Just leftTarget, Just rightTarget) ->
-      leftTarget == rightTarget
-    (Just leftTarget, Nothing) ->
-      integerValueMatchesTarget leftTarget rightInt
-    (Nothing, Just rightTarget) ->
-      integerValueMatchesTarget rightTarget leftInt
-    (Nothing, Nothing) ->
-      True
-
-runtimeFloatMetadataCompatible :: RuntimeFloatMetadata -> RuntimeFloatMetadata -> Bool
-runtimeFloatMetadataCompatible leftMetadata rightMetadata =
-  case (runtimeFloatTargetType leftMetadata, runtimeFloatTargetType rightMetadata) of
-    (Just leftTarget, Just rightTarget) ->
-      leftTarget == rightTarget
-    (Just NumericFloat64, Nothing) ->
-      True
-    (Nothing, Just NumericFloat64) ->
-      True
-    (Nothing, Nothing) ->
-      True
-    _ ->
-      False
 
 liftRuntimeResult :: Monad m => Either Diagnostic value -> ExceptT Diagnostic m value
 liftRuntimeResult result =
@@ -4012,25 +1972,6 @@ lookupDeclaredOperatorCell operatorSymbol env =
             ("operator '" <> operatorSymbol <> "' has no executable binding")
         )
 
-preferredRuntimeMethodCandidates ::
-  Text ->
-  SignaturePayload ->
-  [RuntimeValue] ->
-  [RuntimeMethodCandidate] ->
-  [RuntimeMethodCandidate]
-preferredRuntimeMethodCandidates classParameter methodSignature arguments candidates =
-  case exactMatchingCandidates of
-    [] -> matchingCandidates
-    exactMatches -> exactMatches
-  where
-    exactMatchingCandidates =
-      filter
-        (runtimeMethodCandidateExactlyMatches classParameter methodSignature arguments)
-        matchingCandidates
-    matchingCandidates =
-      filter
-        (runtimeMethodCandidateMatches classParameter methodSignature arguments)
-        candidates
 
 evalValueWithHost ::
   Monad m =>
@@ -4515,15 +2456,11 @@ evalBuiltinWithHost host builtinMode bindingTypeHints builtinFunction arguments 
                 "E3030"
                 ("runtime primitive 'exit!' expects a status in range 0..255, found " <> Text.pack (show status))
             )
-    (BuiltinMap, [mapper, VList elements maybeCollectionTypeHint])
-      | isFunctionValue mapper -> do
-          mappedElements <- traverse (applyRuntimeFunctionWithHost host builtinMode bindingTypeHints mapper) elements
-          let maybeMappedTypeHint = TypeList <$> runtimeMapResultElementType mapper maybeCollectionTypeHint
-          pure (VList mappedElements maybeMappedTypeHint)
-    (BuiltinFilter, [predicate, VList elements maybeTypeHint])
-      | isFunctionValue predicate ->
-          (`VList` maybeTypeHint) <$> filterElementsWithHost host builtinMode bindingTypeHints predicate elements
-    _ -> liftRuntimeResult (evalBuiltin builtinMode bindingTypeHints builtinFunction arguments)
+    _ ->
+      evalBuiltin
+        (applyRuntimeFunctionWithHost host builtinMode bindingTypeHints)
+        builtinFunction
+        arguments
 
 rawHostOutcome :: (success -> RuntimeValue) -> Either HostIOFailure success -> RuntimeValue
 rawHostOutcome renderSuccess outcome =
@@ -4548,26 +2485,6 @@ runtimeHostExitStatus runtimeValue =
     VExplicitResultHints _ innerValue -> runtimeHostExitStatus innerValue
     _ -> Nothing
 
-filterElementsWithHost ::
-  Monad m =>
-  RuntimeHost (RuntimeHostEvaluationT m) ->
-  BuiltinResolutionMode ->
-  Map BindingRuntimeHintKey SignatureType ->
-  RuntimeValue ->
-  [RuntimeValue] ->
-  ExceptT Diagnostic (RuntimeHostEvaluationT m) [RuntimeValue]
-filterElementsWithHost host builtinMode bindingTypeHints predicate values = do
-  results <- traverse applyPredicate values
-  pure [value | (value, True) <- results]
-  where
-    applyPredicate value = do
-      predicateResult <- applyRuntimeFunctionWithHost host builtinMode bindingTypeHints predicate value
-      case predicateResult of
-        VBool shouldKeep -> pure (value, shouldKeep)
-        other ->
-          throwE
-            (runtimeDiagnostic "E3019" ("runtime primitive 'filter' predicate must return Bool, found " <> renderRuntimeType other))
-
 evalBinaryWithHost ::
   Monad m =>
   RuntimeHost (RuntimeHostEvaluationT m) ->
@@ -4578,54 +2495,8 @@ evalBinaryWithHost ::
   RuntimeValue ->
   ExceptT Diagnostic (RuntimeHostEvaluationT m) RuntimeValue
 evalBinaryWithHost host builtinMode bindingTypeHints operatorSymbol leftValue rightValue =
-  case (operatorSymbol, leftValue, rightValue) of
-    ("$", functionValue, argumentValue) ->
-      applyRuntimeFunctionWithHost host builtinMode bindingTypeHints functionValue argumentValue
-    (_, VTyped leftTypeHint leftInnerValue, _)
-      | not (isStrictEqualityOperator operatorSymbol) -> do
-          result <- evalBinaryWithHost host builtinMode bindingTypeHints operatorSymbol leftInnerValue rightValue
-          liftRuntimeResult (preserveLeftTypedNumericOperatorResult operatorSymbol leftTypeHint result)
-    (_, _, VTyped rightTypeHint rightInnerValue)
-      | not (isStrictEqualityOperator operatorSymbol) -> do
-          result <- evalBinaryWithHost host builtinMode bindingTypeHints operatorSymbol leftValue rightInnerValue
-          liftRuntimeResult (preserveRightTypedNumericOperatorResult operatorSymbol leftValue rightTypeHint result)
-    _ -> liftRuntimeResult (evalBinary builtinMode bindingTypeHints operatorSymbol leftValue rightValue)
-
--- | Runtime-specific wrapper for mkDiagnostic.
--- This alias exists solely to improve readability and make it clear that
--- diagnostics are being created in a runtime evaluation context rather than
--- during parsing or type checking.
-runtimeDiagnostic :: Text -> Text -> Diagnostic
-runtimeDiagnostic = mkDiagnostic
-
--- | Render coarse runtime type names for diagnostics.
-renderRuntimeType :: RuntimeValue -> Text
-renderRuntimeType value =
-  case value of
-    VInt _ metadata ->
-      case runtimeIntTargetType metadata of
-        Just targetType -> renderNumericTypeName targetType
-        Nothing -> "Int"
-    VFloat {} -> "Float"
-    VBool {} -> "Bool"
-    VChar {} -> "Char"
-    VText {} -> "Text"
-    VList {} -> "List"
-    VTuple {} -> "Tuple"
-    VSectionLeft {} -> "Function"
-    VSectionRight {} -> "Function"
-    VClosure {} -> "Function"
-    VBuiltin {} -> "Function"
-    VOperator {} -> "Function"
-    VConstructor _ _ _ constructorArguments capturedArgs
-      | constructorIsSaturated constructorArguments capturedArgs -> "Data"
-      | otherwise -> "Function"
-    VQualifiedMethod {} -> "Function"
-    VTyped _ innerValue -> renderRuntimeType innerValue
-    VExplicitTypeApplication _ innerValue -> renderRuntimeType innerValue
-    VExplicitResultHints _ innerValue -> renderRuntimeType innerValue
-    VDeferredHostBinding {} -> "Deferred"
-
-constructorIsSaturated :: [DataConstructorArgument] -> [RuntimeValue] -> Bool
-constructorIsSaturated constructorArguments capturedArgs =
-  length capturedArgs >= length constructorArguments
+  evalBinary
+    (applyRuntimeFunctionWithHost host builtinMode bindingTypeHints)
+    operatorSymbol
+    leftValue
+    rightValue
